refactor(runner): split start test harness
