+++ crates/runner/src/cmd/start/mod.rs
@@ -748,7 +748,7 @@ async fn run(config: RunConfig) -> RunnerResult<()> {
             false
         };
         tokio::select! {
-            // Job discovery via provider (Ably wakeups + HTTP poll).
+            // Job discovery via provider (Ably push + poll).
             // The future is pinned outside the loop so heartbeat/cleanup
             // ticks don't cancel and restart its internal poll timer. See #8747.
             discovered = &mut discover_fut, if can_discover => {
@@ -841,9 +841,9 @@ async fn run(config: RunConfig) -> RunnerResult<()> {
     // -----------------------------------------------------------------------
     let teardown = TeardownTimer::start();
 
-    // Drop the pinned discover future before provider shutdown so any
-    // provider-local discovery resources are released first. This also keeps
-    // the historical shutdown-deadlock regression covered by mock providers.
+    // Drop the pinned discover future so it releases the discovery Mutex.
+    // Without this, provider.shutdown() deadlocks trying to acquire the
+    // same Mutex that the still-alive discover_fut holds.
     drop(discover_fut);
     teardown.event("drop_discover_fut");
 
@@ -998,4996 +998,4 @@ async fn run(config: RunConfig) -> RunnerResult<()> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::job_lifecycle::{
-        ActiveBudgetLease, BudgetOwnership, CompletionPayload, CompletionReady,
-        RunCleanupDisposition, RunCleanupState,
-    };
-    use super::job_spawn::cleanup_panicked_job;
-    use super::sandbox_finalization::{FinalizeContext, finalize_sandbox_for_completion};
-    use super::signals::{
-        LifecycleController, handle_drain_signal, handle_resume_signal, handle_stopping_signal,
-    };
-    use super::*;
-    use std::collections::BTreeMap;
-    use std::fmt;
-    use std::sync::Mutex;
-    use std::sync::atomic::{AtomicUsize, Ordering};
-    use tracing::field::{Field, Visit};
-    use tracing::{Event, Subscriber};
-    use tracing_subscriber::layer::{Context, Layer};
-    use tracing_subscriber::prelude::*;
-
-    use crate::executor;
-    use crate::idle_pool::{
-        IdlePool, IdlePoolConfig, ParkCandidate, ParkCandidateParts, ParkResult, ParkingState,
-    };
-    use crate::resource_budget::BudgetLease;
-    use crate::types::HeartbeatState;
-    use crate::types::{ExecutionContext, SandboxReuseResult};
-    use async_trait::async_trait;
-    use sandbox::SandboxId;
-    use sandbox::{Sandbox, SandboxFactory};
-    use sandbox_mock::{MockSandbox, MockSandboxFactory};
-
-    #[derive(Clone, Default)]
-    struct CapturedEvents {
-        events: Arc<Mutex<Vec<CapturedEvent>>>,
-    }
-
-    #[derive(Clone, Debug, Default)]
-    struct CapturedEvent {
-        fields: BTreeMap<String, String>,
-        field_kinds: BTreeMap<String, &'static str>,
-    }
-
-    impl CapturedEvents {
-        fn entries(&self) -> Vec<CapturedEvent> {
-            self.events.lock().unwrap().clone()
-        }
-
-        fn clear(&self) {
-            self.events.lock().unwrap().clear();
-        }
-    }
-
-    impl<S> Layer<S> for CapturedEvents
-    where
-        S: Subscriber,
-    {
-        fn on_event(&self, event: &Event<'_>, _ctx: Context<'_, S>) {
-            let mut visitor = CapturedEvent::default();
-            event.record(&mut visitor);
-            self.events.lock().unwrap().push(visitor);
-        }
-    }
-
-    impl Visit for CapturedEvent {
-        fn record_str(&mut self, field: &Field, value: &str) {
-            self.fields
-                .insert(field.name().to_string(), value.to_string());
-            self.field_kinds.insert(field.name().to_string(), "str");
-        }
-
-        fn record_u64(&mut self, field: &Field, value: u64) {
-            self.fields
-                .insert(field.name().to_string(), value.to_string());
-            self.field_kinds.insert(field.name().to_string(), "u64");
-        }
-
-        fn record_debug(&mut self, field: &Field, value: &dyn fmt::Debug) {
-            self.fields
-                .insert(field.name().to_string(), format!("{value:?}"));
-            self.field_kinds.insert(field.name().to_string(), "debug");
-        }
-    }
-
-    fn event_with_message<'a>(events: &'a [CapturedEvent], message: &str) -> &'a CapturedEvent {
-        events
-            .iter()
-            .find(|event| {
-                event
-                    .fields
-                    .get("message")
-                    .is_some_and(|value| value == message)
-            })
-            .unwrap_or_else(|| panic!("missing event message {message:?}; captured={events:#?}"))
-    }
-
-    fn register_teardown_timer_callsites() {
-        let timer = TeardownTimer::start();
-        let phase = timer.phase_start("warmup_phase");
-        timer.phase_complete("warmup_phase", phase);
-        timer.event("warmup_event");
-        info!(total_teardown_ms = timer.elapsed_ms(), "runner stopped");
-    }
-
-    #[test]
-    fn teardown_timer_emits_structured_timing_fields() {
-        let captured = CapturedEvents::default();
-        let subscriber = tracing_subscriber::registry().with(captured.clone());
-        let _guard = tracing::subscriber::set_default(subscriber);
-        // These callsites are also hit by parallel runner tests. Reproduce the
-        // problematic shape by registering them on a thread without this
-        // thread-local subscriber, then rebuild the cache for this subscriber.
-        std::thread::spawn(register_teardown_timer_callsites)
-            .join()
-            .expect("callsite warmup thread panicked");
-        tracing::callsite::rebuild_interest_cache();
-        captured.clear();
-
-        let timer = TeardownTimer::start();
-        let phase = timer.phase_start("test_phase");
-        timer.phase_complete("test_phase", phase);
-        timer.event("drop_discover_fut");
-        info!(total_teardown_ms = timer.elapsed_ms(), "runner stopped");
-
-        let events = captured.entries();
-        event_with_message(&events, "teardown started");
-
-        let started = event_with_message(&events, "teardown phase started");
-        assert_eq!(
-            started.fields.get("phase").map(String::as_str),
-            Some("test_phase")
-        );
-        assert!(started.fields.contains_key("elapsed_ms"));
-        assert_eq!(started.field_kinds.get("elapsed_ms").copied(), Some("u64"));
-
-        let complete = event_with_message(&events, "teardown phase complete");
-        assert_eq!(
-            complete.fields.get("phase").map(String::as_str),
-            Some("test_phase")
-        );
-        assert!(complete.fields.contains_key("phase_ms"));
-        assert_eq!(complete.field_kinds.get("phase_ms").copied(), Some("u64"));
-        assert!(complete.fields.contains_key("elapsed_ms"));
-        assert_eq!(complete.field_kinds.get("elapsed_ms").copied(), Some("u64"));
-
-        let event = event_with_message(&events, "teardown phase event");
-        assert_eq!(
-            event.fields.get("phase").map(String::as_str),
-            Some("drop_discover_fut")
-        );
-        assert!(event.fields.contains_key("elapsed_ms"));
-        assert_eq!(event.field_kinds.get("elapsed_ms").copied(), Some("u64"));
-
-        let stopped = event_with_message(&events, "runner stopped");
-        assert!(stopped.fields.contains_key("total_teardown_ms"));
-        assert_eq!(
-            stopped.field_kinds.get("total_teardown_ms").copied(),
-            Some("u64")
-        );
-    }
-
-    fn test_profiles() -> BTreeMap<String, config::ProfileConfig> {
-        let mut m = BTreeMap::new();
-        m.insert(
-            "vm0/default".to_string(),
-            config::ProfileConfig {
-                rootfs_hash: "hash".into(),
-                snapshot_hash: "snap".into(),
-                vcpu: 2,
-                memory_mb: 4096,
-                disk_mb: 10240,
-            },
-        );
-        m
-    }
-
-    fn test_budget_lease() -> (Arc<ResourceBudget>, BudgetLease) {
-        let budget = Arc::new(ResourceBudget::new(8, 32768, 1.0, 0));
-        let lease = ResourceBudget::try_reserve_lease(&budget, 2, 4096).unwrap();
-        (budget, lease)
-    }
-
-    fn test_completion_payload(run_id: RunId, sandbox_id: SandboxId) -> CompletionPayload {
-        CompletionPayload::new(run_id, 0, None, sandbox_id, SandboxReuseResult::PoolMiss)
-    }
-
-    async fn status_active_run_count(path: &std::path::Path) -> usize {
-        let raw = tokio::fs::read_to_string(path).await.unwrap();
-        let status: serde_json::Value = serde_json::from_str(&raw).unwrap();
-        status["active_runs"].as_array().unwrap().len()
-    }
-
-    struct CompletionOrderProvider {
-        budget: Arc<ResourceBudget>,
-        budget_count_at_complete: Arc<AtomicUsize>,
-        active_runs_at_complete: Arc<AtomicUsize>,
-        status_path: std::path::PathBuf,
-    }
-
-    #[async_trait]
-    impl JobProvider for CompletionOrderProvider {
-        async fn discover(&self) -> Option<(RunId, String)> {
-            None
-        }
-
-        async fn claim(&self, _run_id: RunId) -> Option<ExecutionContext> {
-            None
-        }
-
-        async fn complete(
-            &self,
-            _run_id: RunId,
-            _exit_code: i32,
-            _error: Option<&str>,
-            _sandbox_id: Option<SandboxId>,
-            _reuse_result: Option<SandboxReuseResult>,
-        ) {
-            self.budget_count_at_complete
-                .store(self.budget.allocated().2, Ordering::SeqCst);
-            self.active_runs_at_complete.store(
-                status_active_run_count(&self.status_path).await,
-                Ordering::SeqCst,
-            );
-        }
-
-        async fn heartbeat(&self, _state: &HeartbeatState) {}
-
-        async fn shutdown(&self) {}
-    }
-
-    #[tokio::test]
-    async fn completion_ready_complete_and_release_orders_completion_status_and_budget() {
-        let (budget, lease) = test_budget_lease();
-        let budget_count_at_complete = Arc::new(AtomicUsize::new(usize::MAX));
-        let active_runs_at_complete = Arc::new(AtomicUsize::new(usize::MAX));
-        let dir = tempfile::tempdir().unwrap();
-        let status_path = dir.path().join("status.json");
-        let status = StatusTracker::new(status_path.clone(), 4, None, None);
-        let provider = CompletionOrderProvider {
-            budget: Arc::clone(&budget),
-            budget_count_at_complete: Arc::clone(&budget_count_at_complete),
-            active_runs_at_complete: Arc::clone(&active_runs_at_complete),
-            status_path: status_path.clone(),
-        };
-        let cleanup_state = RunCleanupState::new();
-        let run_id = RunId::new_v4();
-        let sandbox_id = SandboxId::new_v4();
-        status.add_run(run_id, sandbox_id).await;
-
-        CompletionReady::new(
-            test_completion_payload(run_id, sandbox_id),
-            BudgetOwnership::active(ActiveBudgetLease::new(lease)),
-        )
-        .complete_and_release(&provider, &status, &cleanup_state)
-        .await;
-
-        assert_eq!(
-            budget_count_at_complete.load(Ordering::SeqCst),
-            1,
-            "active budget must still be held while provider.complete runs",
-        );
-        assert_eq!(
-            active_runs_at_complete.load(Ordering::SeqCst),
-            1,
-            "active status removal must happen after provider.complete",
-        );
-        assert_eq!(
-            status_active_run_count(&status_path).await,
-            0,
-            "active status removal should complete before active budget release returns",
-        );
-        assert_eq!(
-            cleanup_state.disposition(),
-            RunCleanupDisposition::StatusRemoved,
-        );
-        assert_eq!(budget.allocated().2, 0);
-    }
-
-    #[tokio::test]
-    async fn completion_ready_idle_owned_does_not_release_park_candidate_budget() {
-        let (budget, lease) = test_budget_lease();
-        let park_candidate_lease = ActiveBudgetLease::new(lease).into_park_candidate_lease();
-        let budget_count_at_complete = Arc::new(AtomicUsize::new(usize::MAX));
-        let active_runs_at_complete = Arc::new(AtomicUsize::new(usize::MAX));
-        let dir = tempfile::tempdir().unwrap();
-        let status_path = dir.path().join("status.json");
-        let status = StatusTracker::new(status_path.clone(), 4, None, None);
-        let provider = CompletionOrderProvider {
-            budget: Arc::clone(&budget),
-            budget_count_at_complete,
-            active_runs_at_complete,
-            status_path,
-        };
-        let cleanup_state = RunCleanupState::new();
-        let run_id = RunId::new_v4();
-        let sandbox_id = SandboxId::new_v4();
-        status.add_run(run_id, sandbox_id).await;
-
-        CompletionReady::new(
-            test_completion_payload(run_id, sandbox_id),
-            BudgetOwnership::idle_owned(),
-        )
-        .complete_and_release(&provider, &status, &cleanup_state)
-        .await;
-
-        assert_eq!(
-            budget.allocated().2,
-            1,
-            "idle-owned completion must not release the park candidate budget",
-        );
-        assert_eq!(
-            cleanup_state.disposition(),
-            RunCleanupDisposition::StatusRemoved,
-        );
-        drop(park_candidate_lease);
-        assert_eq!(budget.allocated().2, 0);
-    }
-
-    #[tokio::test]
-    async fn completion_ready_does_not_remove_reinserted_active_run() {
-        let (budget, lease) = test_budget_lease();
-        let budget_count_at_complete = Arc::new(AtomicUsize::new(usize::MAX));
-        let active_runs_at_complete = Arc::new(AtomicUsize::new(usize::MAX));
-        let dir = tempfile::tempdir().unwrap();
-        let status_path = dir.path().join("status.json");
-        let status = StatusTracker::new(status_path.clone(), 4, None, None);
-        let provider = CompletionOrderProvider {
-            budget: Arc::clone(&budget),
-            budget_count_at_complete,
-            active_runs_at_complete,
-            status_path: status_path.clone(),
-        };
-        let cleanup_state = RunCleanupState::new();
-        let run_id = RunId::new_v4();
-        let completed_sandbox_id = SandboxId::new_v4();
-        let current_sandbox_id = SandboxId::new_v4();
-        status.add_run(run_id, completed_sandbox_id).await;
-        status.add_run(run_id, current_sandbox_id).await;
-
-        CompletionReady::new(
-            test_completion_payload(run_id, completed_sandbox_id),
-            BudgetOwnership::active(ActiveBudgetLease::new(lease)),
-        )
-        .complete_and_release(&provider, &status, &cleanup_state)
-        .await;
-
-        assert_eq!(
-            status_active_run_records(&status_path).await,
-            vec![(run_id.to_string(), current_sandbox_id.to_string())],
-        );
-        assert_eq!(
-            cleanup_state.disposition(),
-            RunCleanupDisposition::StatusRemoved,
-        );
-        assert_eq!(budget.allocated().2, 0);
-    }
-
-    #[tokio::test]
-    async fn rejected_park_budget_is_recovered_as_active_and_released_after_completion() {
-        let (budget, lease) = test_budget_lease();
-        let budget_count_at_complete = Arc::new(AtomicUsize::new(usize::MAX));
-        let active_runs_at_complete = Arc::new(AtomicUsize::new(usize::MAX));
-        let dir = tempfile::tempdir().unwrap();
-        let status_path = dir.path().join("status.json");
-        let status = StatusTracker::new(status_path.clone(), 4, None, None);
-        let provider = CompletionOrderProvider {
-            budget: Arc::clone(&budget),
-            budget_count_at_complete: Arc::clone(&budget_count_at_complete),
-            active_runs_at_complete,
-            status_path,
-        };
-        let cleanup_state = RunCleanupState::new();
-        let run_id = RunId::new_v4();
-        let sandbox_id = SandboxId::new_v4();
-        status.add_run(run_id, sandbox_id).await;
-
-        CompletionReady::new(
-            test_completion_payload(run_id, sandbox_id),
-            BudgetOwnership::active(ActiveBudgetLease::from_rejected_park(lease)),
-        )
-        .complete_and_release(&provider, &status, &cleanup_state)
-        .await;
-
-        assert_eq!(
-            budget_count_at_complete.load(Ordering::SeqCst),
-            1,
-            "rejected park must retain active budget through provider.complete",
-        );
-        assert_eq!(budget.allocated().2, 0);
-    }
-
-    #[tokio::test]
-    async fn finalizer_closes_network_log_session_before_parking() {
-        let (_budget, lease) = test_budget_lease();
-        let dir = tempfile::tempdir().unwrap();
-        let status = Arc::new(StatusTracker::new(
-            dir.path().join("status.json"),
-            4,
-            None,
-            None,
-        ));
-        status.write_initial().await;
-        let parking_gate = ParkingGate::new_open();
-        let idle_pool: SharedIdlePool =
-            Arc::new(tokio::sync::Mutex::new(IdlePool::new_with_parking_gate(
-                IdlePoolConfig {
-                    default_timeout: Duration::from_secs(300),
-                    max_idle: 10,
-                },
-                parking_gate.clone(),
-            )));
-        let network_log_manager = NetworkLogManager::new();
-        let network_log_path = dir.path().join("network.jsonl");
-        let network_log_session = network_log_manager
-            .register_source_ip("10.0.0.1", network_log_path)
-            .await;
-        let run_id = RunId::new_v4();
-        let sandbox_id = SandboxId::new_v4();
-
-        let _completion_ready = finalize_sandbox_for_completion(
-            Some(Box::new(MockSandbox::new("network-log-park"))),
-            ActiveBudgetLease::new(lease),
-            CompletionPayload::new(run_id, 0, None, sandbox_id, SandboxReuseResult::PoolMiss),
-            FinalizeContext {
-                run_id,
-                sandbox_id,
-                profile_name: "vm0/default".into(),
-                session_id: Some("sess-network-log-park".into()),
-                guest_session_id: None,
-                source_ip: "10.0.0.1".into(),
-                network_log_session: Some(network_log_session),
-                storage_fingerprints: crate::idle_pool::StorageFingerprints::default(),
-                factory: Arc::new(Box::new(MockSandboxFactory::new()) as Box<dyn SandboxFactory>),
-                idle_pool: Arc::clone(&idle_pool),
-                status,
-                park_notify: Arc::new(tokio::sync::Notify::new()),
-                parking_gate,
-                network_log_drain: NetworkLogDrainCoordinator::noop(),
-                exit_code: 0,
-                cancel: CancellationToken::new(),
-                cleanup_state: RunCleanupState::new(),
-                outer_job_panic: None,
-            },
-        )
-        .await;
-
-        assert_eq!(idle_pool.lock().await.len(), 1);
-        assert!(
-            !network_log_manager
-                .append_for_ip(
-                    "10.0.0.1",
-                    serde_json::json!({"type":"dns","host":"after-park.test"})
-                )
-                .await,
-            "parked sandbox must not retain the previous run's network-log attribution",
-        );
-    }
-
-    #[tokio::test]
-    async fn finalizer_closes_network_log_session_before_cancel_destroy() {
-        let (_budget, lease) = test_budget_lease();
-        let dir = tempfile::tempdir().unwrap();
-        let status = Arc::new(StatusTracker::new(
-            dir.path().join("status.json"),
-            4,
-            None,
-            None,
-        ));
-        status.write_initial().await;
-        let parking_gate = ParkingGate::new_open();
-        let idle_pool: SharedIdlePool =
-            Arc::new(tokio::sync::Mutex::new(IdlePool::new_with_parking_gate(
-                IdlePoolConfig {
-                    default_timeout: Duration::from_secs(300),
-                    max_idle: 10,
-                },
-                parking_gate.clone(),
-            )));
-        let network_log_manager = NetworkLogManager::new();
-        let network_log_path = dir.path().join("network.jsonl");
-        let network_log_session = network_log_manager
-            .register_source_ip("10.0.0.1", network_log_path)
-            .await;
-        let cancel = CancellationToken::new();
-        cancel.cancel();
-        let run_id = RunId::new_v4();
-        let sandbox_id = SandboxId::new_v4();
-
-        let _completion_ready = finalize_sandbox_for_completion(
-            Some(Box::new(MockSandbox::new("network-log-cancel"))),
-            ActiveBudgetLease::new(lease),
-            CompletionPayload::new(run_id, 0, None, sandbox_id, SandboxReuseResult::PoolMiss),
-            FinalizeContext {
-                run_id,
-                sandbox_id,
-                profile_name: "vm0/default".into(),
-                session_id: Some("sess-network-log-cancel".into()),
-                guest_session_id: None,
-                source_ip: "10.0.0.1".into(),
-                network_log_session: Some(network_log_session),
-                storage_fingerprints: crate::idle_pool::StorageFingerprints::default(),
-                factory: Arc::new(Box::new(MockSandboxFactory::new()) as Box<dyn SandboxFactory>),
-                idle_pool: Arc::clone(&idle_pool),
-                status,
-                park_notify: Arc::new(tokio::sync::Notify::new()),
-                parking_gate,
-                network_log_drain: NetworkLogDrainCoordinator::noop(),
-                exit_code: 0,
-                cancel,
-                cleanup_state: RunCleanupState::new(),
-                outer_job_panic: None,
-            },
-        )
-        .await;
-
-        assert_eq!(idle_pool.lock().await.len(), 0);
-        assert!(
-            !network_log_manager
-                .append_for_ip(
-                    "10.0.0.1",
-                    serde_json::json!({"type":"dns","host":"after-destroy.test"})
-                )
-                .await,
-            "cancelled destroyed sandbox must not retain network-log attribution",
-        );
-    }
-
-    #[test]
-    fn active_budget_drop_releases_budget_as_raii_fallback() {
-        let (budget, lease) = test_budget_lease();
-        drop(ActiveBudgetLease::new(lease));
-        assert_eq!(budget.allocated().2, 0);
-    }
-
-    #[test]
-    fn run_cleanup_state_does_not_downgrade_precise_ownership() {
-        let state = RunCleanupState::new();
-
-        state.mark_idle_pool_owned();
-        state.mark_destroy_completed();
-        assert_eq!(state.disposition(), RunCleanupDisposition::IdlePoolOwned);
-
-        state.mark_status_removed();
-        state.mark_idle_pool_owned();
-        assert_eq!(state.disposition(), RunCleanupDisposition::StatusRemoved);
-    }
-
-    #[tokio::test]
-    async fn panic_cleanup_status_removed_only_clears_cancel_token() {
-        let dir = tempfile::tempdir().unwrap();
-        let status_path = dir.path().join("status.json");
-        let status = Arc::new(StatusTracker::new(status_path.clone(), 4, None, None));
-        let idle_pool: SharedIdlePool =
-            Arc::new(tokio::sync::Mutex::new(IdlePool::new(IdlePoolConfig {
-                default_timeout: Duration::from_secs(300),
-                max_idle: 10,
-            })));
-        let tokens: Arc<tokio::sync::Mutex<HashMap<RunId, CancellationToken>>> =
-            Arc::new(tokio::sync::Mutex::new(HashMap::new()));
-        let orphans = OrphanedActiveRuns::new();
-        let cleanup_state = RunCleanupState::new();
-        let run_id = RunId::new_v4();
-        let sandbox_id = SandboxId::new_v4();
-        status.add_run(run_id, sandbox_id).await;
-        status.remove_run_if_matching(run_id, sandbox_id).await;
-        tokens.lock().await.insert(run_id, CancellationToken::new());
-        cleanup_state.mark_status_removed();
-
-        cleanup_panicked_job(
-            run_id,
-            sandbox_id,
-            Arc::clone(&tokens),
-            Arc::clone(&status),
-            idle_pool,
-            cleanup_state,
-            orphans.clone(),
-        )
-        .await;
-
-        assert!(!tokens.lock().await.contains_key(&run_id));
-        let (_idle_sessions, active_runs) =
-            status_idle_sessions_and_active_runs(&status_path).await;
-        assert!(active_runs.is_empty());
-        assert_eq!(orphans.len().await, 0);
-    }
-
-    #[tokio::test]
-    async fn panic_cleanup_active_unknown_keeps_active_and_registers_orphan() {
-        let dir = tempfile::tempdir().unwrap();
-        let status_path = dir.path().join("status.json");
-        let status = Arc::new(StatusTracker::new(status_path.clone(), 4, None, None));
-        let idle_pool: SharedIdlePool =
-            Arc::new(tokio::sync::Mutex::new(IdlePool::new(IdlePoolConfig {
-                default_timeout: Duration::from_secs(300),
-                max_idle: 10,
-            })));
-        let tokens: Arc<tokio::sync::Mutex<HashMap<RunId, CancellationToken>>> =
-            Arc::new(tokio::sync::Mutex::new(HashMap::new()));
-        let orphans = OrphanedActiveRuns::new();
-        let run_id = RunId::new_v4();
-        let sandbox_id = SandboxId::new_v4();
-        status.add_run(run_id, sandbox_id).await;
-        tokens.lock().await.insert(run_id, CancellationToken::new());
-
-        cleanup_panicked_job(
-            run_id,
-            sandbox_id,
-            Arc::clone(&tokens),
-            Arc::clone(&status),
-            idle_pool,
-            RunCleanupState::new(),
-            orphans.clone(),
-        )
-        .await;
-
-        assert!(!tokens.lock().await.contains_key(&run_id));
-        let (_idle_sessions, active_runs) =
-            status_idle_sessions_and_active_runs(&status_path).await;
-        assert_eq!(active_runs, vec![run_id.to_string()]);
-        assert_eq!(orphans.len().await, 1);
-    }
-
-    #[tokio::test]
-    async fn panic_cleanup_destroy_completed_removes_active_run() {
-        let dir = tempfile::tempdir().unwrap();
-        let status_path = dir.path().join("status.json");
-        let status = Arc::new(StatusTracker::new(status_path.clone(), 4, None, None));
-        let idle_pool: SharedIdlePool =
-            Arc::new(tokio::sync::Mutex::new(IdlePool::new(IdlePoolConfig {
-                default_timeout: Duration::from_secs(300),
-                max_idle: 10,
-            })));
-        let tokens: Arc<tokio::sync::Mutex<HashMap<RunId, CancellationToken>>> =
-            Arc::new(tokio::sync::Mutex::new(HashMap::new()));
-        let orphans = OrphanedActiveRuns::new();
-        let cleanup_state = RunCleanupState::new();
-        let run_id = RunId::new_v4();
-        let sandbox_id = SandboxId::new_v4();
-        status.add_run(run_id, sandbox_id).await;
-        tokens.lock().await.insert(run_id, CancellationToken::new());
-        cleanup_state.mark_destroy_completed();
-
-        cleanup_panicked_job(
-            run_id,
-            sandbox_id,
-            Arc::clone(&tokens),
-            Arc::clone(&status),
-            idle_pool,
-            cleanup_state,
-            orphans.clone(),
-        )
-        .await;
-
-        assert!(!tokens.lock().await.contains_key(&run_id));
-        let (_idle_sessions, active_runs) =
-            status_idle_sessions_and_active_runs(&status_path).await;
-        assert!(active_runs.is_empty());
-        assert_eq!(orphans.len().await, 0);
-    }
-
-    #[tokio::test]
-    async fn panic_cleanup_destroy_completed_does_not_remove_reinserted_active_run() {
-        let dir = tempfile::tempdir().unwrap();
-        let status_path = dir.path().join("status.json");
-        let status = Arc::new(StatusTracker::new(status_path.clone(), 4, None, None));
-        let idle_pool: SharedIdlePool =
-            Arc::new(tokio::sync::Mutex::new(IdlePool::new(IdlePoolConfig {
-                default_timeout: Duration::from_secs(300),
-                max_idle: 10,
-            })));
-        let tokens: Arc<tokio::sync::Mutex<HashMap<RunId, CancellationToken>>> =
-            Arc::new(tokio::sync::Mutex::new(HashMap::new()));
-        let orphans = OrphanedActiveRuns::new();
-        let cleanup_state = RunCleanupState::new();
-        let run_id = RunId::new_v4();
-        let completed_sandbox_id = SandboxId::new_v4();
-        let current_sandbox_id = SandboxId::new_v4();
-        status.add_run(run_id, completed_sandbox_id).await;
-        status.add_run(run_id, current_sandbox_id).await;
-        tokens.lock().await.insert(run_id, CancellationToken::new());
-        cleanup_state.mark_destroy_completed();
-
-        cleanup_panicked_job(
-            run_id,
-            completed_sandbox_id,
-            Arc::clone(&tokens),
-            Arc::clone(&status),
-            idle_pool,
-            cleanup_state,
-            orphans.clone(),
-        )
-        .await;
-
-        assert!(!tokens.lock().await.contains_key(&run_id));
-        assert_eq!(
-            status_active_run_records(&status_path).await,
-            vec![(run_id.to_string(), current_sandbox_id.to_string())],
-        );
-        assert_eq!(orphans.len().await, 0);
-    }
-
-    #[tokio::test]
-    async fn panic_cleanup_idle_pool_owned_refreshes_idle_status_before_removing_active() {
-        let dir = tempfile::tempdir().unwrap();
-        let status_path = dir.path().join("status.json");
-        let status = Arc::new(StatusTracker::new(status_path.clone(), 4, None, None));
-        let idle_pool: SharedIdlePool =
-            Arc::new(tokio::sync::Mutex::new(IdlePool::new(IdlePoolConfig {
-                default_timeout: Duration::from_secs(300),
-                max_idle: 10,
-            })));
-        let tokens: Arc<tokio::sync::Mutex<HashMap<RunId, CancellationToken>>> =
-            Arc::new(tokio::sync::Mutex::new(HashMap::new()));
-        let orphans = OrphanedActiveRuns::new();
-        let cleanup_state = RunCleanupState::new();
-        let run_id = RunId::new_v4();
-        let sandbox_id = SandboxId::new_v4();
-        let budget = Arc::new(ResourceBudget::new(2, 4096, 1.0, 0));
-        let lease = ResourceBudget::try_reserve_lease(&budget, 2, 4096).unwrap();
-        let candidate = ParkCandidate::from_parked_parts(ParkCandidateParts {
-            sandbox: Box::new(MockSandbox::new("idle-owned-cleanup")),
-            factory: Arc::new(Box::new(MockSandboxFactory::new()) as Box<dyn SandboxFactory>),
-            session_id: "sess-idle-owned-cleanup".into(),
-            sandbox_id,
-            profile_name: "vm0/default".into(),
-            budget_lease: lease,
-            source_ip: "10.0.0.1".into(),
-            storage_fingerprints: crate::idle_pool::StorageFingerprints::default(),
-        });
-        assert!(matches!(
-            idle_pool.lock().await.park(candidate),
-            ParkResult::Parked
-        ));
-        status.add_run(run_id, sandbox_id).await;
-        tokens.lock().await.insert(run_id, CancellationToken::new());
-        cleanup_state.mark_idle_pool_owned();
-
-        cleanup_panicked_job(
-            run_id,
-            sandbox_id,
-            Arc::clone(&tokens),
-            Arc::clone(&status),
-            Arc::clone(&idle_pool),
-            cleanup_state,
-            orphans.clone(),
-        )
-        .await;
-
-        assert!(!tokens.lock().await.contains_key(&run_id));
-        let (idle_sessions, active_runs) = status_idle_sessions_and_active_runs(&status_path).await;
-        assert_eq!(idle_sessions, vec!["sess-idle-owned-cleanup"]);
-        assert!(active_runs.is_empty());
-        assert_eq!(orphans.len().await, 0);
-    }
-
-    struct PanickingDestroyFactory;
-
-    #[async_trait]
-    impl SandboxFactory for PanickingDestroyFactory {
-        fn name(&self) -> &str {
-            "panic-destroy"
-        }
-
-        fn config_hash(&self) -> String {
-            "panic-destroy".into()
-        }
-
-        async fn startup(&mut self) -> sandbox::Result<()> {
-            Ok(())
-        }
-
-        async fn create(
-            &self,
-            config: sandbox::SandboxConfig,
-        ) -> sandbox::Result<Box<dyn Sandbox>> {
-            Ok(Box::new(MockSandbox::new(config.id.to_string())))
-        }
-
-        async fn destroy(&self, _sandbox: Box<dyn Sandbox>) {
-            panic!("simulated destroy panic");
-        }
-
-        async fn shutdown(&mut self) {}
-    }
-
-    struct PanickingDestroyRuntime;
-
-    #[async_trait]
-    impl SandboxRuntime for PanickingDestroyRuntime {
-        async fn create_factory(
-            &self,
-            _config: sandbox::FactoryConfig,
-        ) -> sandbox::Result<Box<dyn SandboxFactory>> {
-            Ok(Box::new(PanickingDestroyFactory))
-        }
-
-        async fn shutdown(&mut self) {}
-    }
-
-    struct FailingCreateFactory;
-
-    #[async_trait]
-    impl SandboxFactory for FailingCreateFactory {
-        fn name(&self) -> &str {
-            "failing-create"
-        }
-
-        fn config_hash(&self) -> String {
-            "failing-create".into()
-        }
-
-        async fn startup(&mut self) -> sandbox::Result<()> {
-            Ok(())
-        }
-
-        async fn create(
-            &self,
-            _config: sandbox::SandboxConfig,
-        ) -> sandbox::Result<Box<dyn Sandbox>> {
-            Err(sandbox::SandboxError::Initialization {
-                phase: sandbox::SandboxInitializationPhase::SandboxAllocation,
-                message: "create failed".into(),
-            })
-        }
-
-        async fn destroy(&self, _sandbox: Box<dyn Sandbox>) {}
-
-        async fn shutdown(&mut self) {}
-    }
-
-    struct FailingCreateRuntime;
-
-    #[async_trait]
-    impl SandboxRuntime for FailingCreateRuntime {
-        async fn create_factory(
-            &self,
-            _config: sandbox::FactoryConfig,
-        ) -> sandbox::Result<Box<dyn SandboxFactory>> {
-            Ok(Box::new(FailingCreateFactory))
-        }
-
-        async fn shutdown(&mut self) {}
-    }
-
-    struct RecordingDestroyFactory {
-        destroy_count: Arc<AtomicUsize>,
-    }
-
-    #[async_trait]
-    impl SandboxFactory for RecordingDestroyFactory {
-        fn name(&self) -> &str {
-            "recording-destroy"
-        }
-
-        fn config_hash(&self) -> String {
-            "recording-destroy".into()
-        }
-
-        async fn startup(&mut self) -> sandbox::Result<()> {
-            Ok(())
-        }
-
-        async fn create(
-            &self,
-            config: sandbox::SandboxConfig,
-        ) -> sandbox::Result<Box<dyn Sandbox>> {
-            Ok(Box::new(MockSandbox::new(config.id.to_string())))
-        }
-
-        async fn destroy(&self, _sandbox: Box<dyn Sandbox>) {
-            self.destroy_count.fetch_add(1, Ordering::SeqCst);
-        }
-
-        async fn shutdown(&mut self) {}
-    }
-
-    #[derive(Clone)]
-    struct BlockingDestroyHooks {
-        entered: Arc<tokio::sync::Notify>,
-        release: Arc<tokio::sync::Notify>,
-        destroy_count: Arc<AtomicUsize>,
-    }
-
-    impl BlockingDestroyHooks {
-        fn new() -> Self {
-            Self {
-                entered: Arc::new(tokio::sync::Notify::new()),
-                release: Arc::new(tokio::sync::Notify::new()),
-                destroy_count: Arc::new(AtomicUsize::new(0)),
-            }
-        }
-    }
-
-    struct BlockingDestroyRuntime {
-        hooks: BlockingDestroyHooks,
-    }
-
-    #[async_trait]
-    impl SandboxRuntime for BlockingDestroyRuntime {
-        async fn create_factory(
-            &self,
-            _config: sandbox::FactoryConfig,
-        ) -> sandbox::Result<Box<dyn SandboxFactory>> {
-            Ok(Box::new(BlockingDestroyFactory {
-                hooks: self.hooks.clone(),
-            }))
-        }
-
-        async fn shutdown(&mut self) {}
-    }
-
-    struct BlockingDestroyFactory {
-        hooks: BlockingDestroyHooks,
-    }
-
-    #[async_trait]
-    impl SandboxFactory for BlockingDestroyFactory {
-        fn name(&self) -> &str {
-            "blocking-destroy"
-        }
-
-        fn config_hash(&self) -> String {
-            "blocking-destroy".into()
-        }
-
-        async fn startup(&mut self) -> sandbox::Result<()> {
-            Ok(())
-        }
-
-        async fn create(
-            &self,
-            config: sandbox::SandboxConfig,
-        ) -> sandbox::Result<Box<dyn Sandbox>> {
-            Ok(Box::new(MockSandbox::new(config.id.to_string())))
-        }
-
-        async fn destroy(&self, _sandbox: Box<dyn Sandbox>) {
-            self.hooks.destroy_count.fetch_add(1, Ordering::SeqCst);
-            self.hooks.entered.notify_waiters();
-            self.hooks.release.notified().await;
-        }
-
-        async fn shutdown(&mut self) {}
-    }
-
-    #[tokio::test]
-    async fn idle_destroy_panic_releases_budget_lease() {
-        let budget = Arc::new(ResourceBudget::new(2, 4096, 1.0, 0));
-        let lease = ResourceBudget::try_reserve_lease(&budget, 2, 4096).unwrap();
-        let candidate = ParkCandidate::from_parked_parts(ParkCandidateParts {
-            sandbox: Box::new(MockSandbox::new("panic-destroy")),
-            factory: Arc::new(Box::new(PanickingDestroyFactory) as Box<dyn SandboxFactory>),
-            session_id: "sess-panic".into(),
-            sandbox_id: SandboxId::new_v4(),
-            profile_name: "vm0/default".into(),
-            budget_lease: lease,
-            source_ip: "10.0.0.1".into(),
-            storage_fingerprints: crate::idle_pool::StorageFingerprints::default(),
-        });
-        let mut pool = IdlePool::new(IdlePoolConfig {
-            default_timeout: Duration::from_secs(300),
-            max_idle: 0,
-        });
-        assert!(matches!(pool.park(candidate), ParkResult::Parked));
-        let jobs = pool.drain();
-
-        destroy_idle_jobs_and_wait(jobs, "test_destroy_panic").await;
-
-        assert_eq!(budget.allocated(), (0, 0, 0));
-    }
-
-    #[tokio::test]
-    async fn idle_stop_panic_still_attempts_destroy_and_releases_budget_lease() {
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::new());
-        overrides.push_stop_panic("simulated idle stop panic");
-        let sandbox_factory = MockSandboxFactory::with_overrides(overrides);
-        let sandbox = sandbox_factory
-            .create(sandbox::SandboxConfig {
-                id: SandboxId::new_v4(),
-                resources: sandbox::ResourceLimits {
-                    cpu_count: 2,
-                    memory_mb: 4096,
-                },
-            })
-            .await
-            .expect("create sandbox");
-
-        let budget = Arc::new(ResourceBudget::new(2, 4096, 1.0, 0));
-        let lease = ResourceBudget::try_reserve_lease(&budget, 2, 4096).unwrap();
-        let destroy_count = Arc::new(AtomicUsize::new(0));
-        let candidate = ParkCandidate::from_parked_parts(ParkCandidateParts {
-            sandbox,
-            factory: Arc::new(Box::new(RecordingDestroyFactory {
-                destroy_count: Arc::clone(&destroy_count),
-            }) as Box<dyn SandboxFactory>),
-            session_id: "sess-stop-panic".into(),
-            sandbox_id: SandboxId::new_v4(),
-            profile_name: "vm0/default".into(),
-            budget_lease: lease,
-            source_ip: "10.0.0.1".into(),
-            storage_fingerprints: crate::idle_pool::StorageFingerprints::default(),
-        });
-        let mut pool = IdlePool::new(IdlePoolConfig {
-            default_timeout: Duration::from_secs(300),
-            max_idle: 0,
-        });
-        assert!(matches!(pool.park(candidate), ParkResult::Parked));
-        let mut jobs = pool.drain();
-        assert_eq!(jobs.len(), 1);
-
-        jobs.pop().unwrap().run().await;
-
-        assert_eq!(destroy_count.load(Ordering::SeqCst), 1);
-        assert_eq!(budget.allocated(), (0, 0, 0));
-    }
-
-    // =======================================================================
-    // Main loop integration tests
-    // =======================================================================
-
-    use crate::provider::mock::{MockJobProvider, MockProviderHandle};
-    use sandbox_mock::MockSandboxRuntime;
-
-    /// Everything a test needs to drive the main loop.
-    struct MockRunEnv {
-        handle: MockProviderHandle,
-        provider: Arc<MockJobProvider>,
-        idle_pool: SharedIdlePool,
-        lifecycle: LifecycleController,
-        parking_gate: ParkingGate,
-        mode_tx: tokio::sync::watch::Sender<RunnerMode>,
-        cancel_tokens: Arc<tokio::sync::Mutex<HashMap<RunId, CancellationToken>>>,
-        cancel: CancellationToken,
-        _temp_dir: tempfile::TempDir,
-    }
-
-    impl MockRunEnv {
-        /// Simulate SIGUSR1 by driving the real `handle_drain_signal` so
-        /// tests exercise the same state-guard path production does
-        /// (ignored unless current mode is Running).
-        fn drain(&self) {
-            handle_drain_signal(&self.lifecycle);
-        }
-
-        /// Simulate SIGUSR2 via the real `handle_resume_signal` — only
-        /// transitions when current mode is Draining.
-        fn resume(&self) {
-            handle_resume_signal(&self.lifecycle);
-        }
-
-        /// Simulate SIGTERM by driving the real `handle_stopping_signal`.
-        /// Keeps the test path in sync with production.
-        async fn trigger_stopping(&self) {
-            handle_stopping_signal("TEST", &self.cancel, &self.cancel_tokens, &self.lifecycle)
-                .await;
-        }
-    }
-
-    /// Assemble a complete `RunConfig` with all mock/noop dependencies.
-    fn mock_run_config(
-        profiles: BTreeMap<String, config::ProfileConfig>,
-        budget_vcpu: u32,
-        budget_memory_mb: u32,
-        max_concurrent: usize,
-    ) -> (RunConfig, MockRunEnv) {
-        build_mock_run_config(
-            profiles,
-            budget_vcpu,
-            budget_memory_mb,
-            max_concurrent,
-            MockJobProvider::new,
-        )
-    }
-
-    /// Variant with an explicit poll delay for regression testing.
-    fn mock_run_config_with_delay(
-        profiles: BTreeMap<String, config::ProfileConfig>,
-        budget_vcpu: u32,
-        budget_memory_mb: u32,
-        max_concurrent: usize,
-        poll_delay: Duration,
-    ) -> (RunConfig, MockRunEnv) {
-        build_mock_run_config(
-            profiles,
-            budget_vcpu,
-            budget_memory_mb,
-            max_concurrent,
-            |cancel| MockJobProvider::with_poll_delay(cancel, Some(poll_delay)),
-        )
-    }
-
-    fn build_mock_run_config(
-        profiles: BTreeMap<String, config::ProfileConfig>,
-        budget_vcpu: u32,
-        budget_memory_mb: u32,
-        max_concurrent: usize,
-        make_provider: impl FnOnce(CancellationToken) -> (Arc<MockJobProvider>, MockProviderHandle),
-    ) -> (RunConfig, MockRunEnv) {
-        build_mock_run_config_with_runtime(
-            profiles,
-            budget_vcpu,
-            budget_memory_mb,
-            max_concurrent,
-            make_provider,
-            Box::new(MockSandboxRuntime::new()),
-            "http://localhost:0",
-        )
-    }
-
-    /// Variant that points the runner's HTTP client at an explicit URL. Used by
-    /// tests that spin up an `httpmock::MockServer` to observe webhook traffic.
-    fn mock_run_config_with_api_url(
-        profiles: BTreeMap<String, config::ProfileConfig>,
-        budget_vcpu: u32,
-        budget_memory_mb: u32,
-        max_concurrent: usize,
-        api_url: &str,
-    ) -> (RunConfig, MockRunEnv) {
-        build_mock_run_config_with_runtime(
-            profiles,
-            budget_vcpu,
-            budget_memory_mb,
-            max_concurrent,
-            MockJobProvider::new,
-            Box::new(MockSandboxRuntime::new()),
-            api_url,
-        )
-    }
-
-    fn build_mock_run_config_with_runtime(
-        profiles: BTreeMap<String, config::ProfileConfig>,
-        budget_vcpu: u32,
-        budget_memory_mb: u32,
-        max_concurrent: usize,
-        make_provider: impl FnOnce(CancellationToken) -> (Arc<MockJobProvider>, MockProviderHandle),
-        runtime: Box<dyn sandbox::SandboxRuntime>,
-        api_url: &str,
-    ) -> (RunConfig, MockRunEnv) {
-        let temp_dir = tempfile::tempdir().unwrap();
-        let cancel = CancellationToken::new();
-        let (provider, handle) = make_provider(cancel.clone());
-        let provider_ref = Arc::clone(&provider);
-
-        let (mode_tx, mode_rx) = tokio::sync::watch::channel(RunnerMode::Running);
-        let parking_gate = ParkingGate::new_open();
-        let lifecycle = LifecycleController::new(mode_tx, parking_gate.clone());
-        let cancel_tokens: Arc<tokio::sync::Mutex<HashMap<RunId, CancellationToken>>> =
-            Arc::new(tokio::sync::Mutex::new(HashMap::new()));
-
-        let home = HomePaths::with_root(temp_dir.path().to_path_buf());
-        let registry_path = temp_dir.path().join("registry.json");
-        let lock_path = temp_dir.path().join("registry.lock");
-        // Write empty registry file so ProxyRegistryHandle can read it.
-        std::fs::write(&registry_path, r#"{"vms":{},"updatedAt":0}"#).unwrap();
-        let registry = proxy::ProxyRegistryHandle::new(registry_path, lock_path);
-
-        let log_dir = temp_dir.path().join("logs");
-        std::fs::create_dir_all(&log_dir).unwrap();
-
-        let (mitm, mitm_crash_rx) = proxy::MitmProxy::noop();
-        let min_vcpu = profiles_min_vcpu(&profiles);
-        let min_memory_mb = profiles_min_memory(&profiles);
-        let idle_pool: SharedIdlePool =
-            Arc::new(tokio::sync::Mutex::new(IdlePool::new_with_parking_gate(
-                IdlePoolConfig {
-                    default_timeout: Duration::from_secs(300),
-                    max_idle: 10,
-                },
-                parking_gate.clone(),
-            )));
-
-        let config = RunConfig {
-            id: "test-runner".into(),
-            name: "test".into(),
-            group: "test-group".into(),
-            profiles,
-            runtime,
-            home: home.clone(),
-            budget: Arc::new(ResourceBudget::new(
-                budget_vcpu,
-                budget_memory_mb,
-                1.0,
-                max_concurrent,
-            )),
-            idle_pool: Arc::clone(&idle_pool),
-            parking_gate: parking_gate.clone(),
-            status: Arc::new(StatusTracker::new(
-                temp_dir.path().join("status.json"),
-                max_concurrent,
-                None,
-                None,
-            )),
-            mitm,
-            mitm_crash_rx,
-            provider,
-            cancel_tokens: Arc::clone(&cancel_tokens),
-            cancel: cancel.clone(),
-            exec_config: Arc::new(executor::ExecutorConfig {
-                api_url: api_url.to_string(),
-                registry,
-                http: crate::http::HttpClient::new(api_url.to_string()).unwrap(),
-                log_paths: crate::paths::LogPaths::new(log_dir),
-                network_log_manager: NetworkLogManager::new(),
-                network_log_drain: NetworkLogDrainCoordinator::noop(),
-                home,
-            }),
-            firecracker: config::FirecrackerConfig {
-                binary: PathBuf::new(),
-                kernel: PathBuf::new(),
-            },
-            base_dir: temp_dir.path().to_path_buf(),
-            min_vcpu,
-            min_memory_mb,
-            kmsg_handle: kmsg_log::KmsgHandle::noop(),
-            dns_handle: crate::dns::DnsProxy::noop(),
-            orphan_reap_process_discovery: None,
-            signal_source: SignalSource::Override(SignalController {
-                mode_rx,
-                lifecycle: lifecycle.clone(),
-                handler_abort: None,
-            }),
-            outer_job_panic: None,
-        };
-
-        let env = MockRunEnv {
-            handle,
-            provider: provider_ref,
-            idle_pool,
-            lifecycle: lifecycle.clone(),
-            parking_gate,
-            mode_tx: lifecycle.mode_tx().clone(),
-            cancel_tokens,
-            cancel,
-            _temp_dir: temp_dir,
-        };
-        (config, env)
-    }
-
-    fn profiles_min_vcpu(profiles: &BTreeMap<String, config::ProfileConfig>) -> u32 {
-        profiles.values().map(|p| p.vcpu).min().unwrap_or(1)
-    }
-
-    fn profiles_min_memory(profiles: &BTreeMap<String, config::ProfileConfig>) -> u32 {
-        profiles.values().map(|p| p.memory_mb).min().unwrap_or(1)
-    }
-
-    fn minimal_context(run_id: RunId) -> crate::types::ExecutionContext {
-        crate::types::ExecutionContext {
-            run_id,
-            prompt: "test".into(),
-            append_system_prompt: None,
-            _agent_compose_version_id: None,
-            vars: None,
-            checkpoint_id: None,
-            sandbox_token: "tok".into(),
-            working_dir: "/workspace".into(),
-            storage_manifest: None,
-            environment: None,
-            resume_session: None,
-            secret_values: None,
-            encrypted_secrets: None,
-            secret_connector_map: None,
-            cli_agent_type: String::new(),
-            debug_no_mock_claude: None,
-            debug_no_mock_codex: None,
-            api_start_time: None,
-            user_timezone: None,
-            capture_network_bodies: None,
-            firewalls: None,
-            network_policies: None,
-            disallowed_tools: None,
-            tools: None,
-            settings: None,
-            experimental_profile: None,
-            feature_flags: None,
-            billable_firewalls: vec![],
-            model_usage_provider: None,
-        }
-    }
-
-    /// Push a job to the mock provider and pre-configure its claim result.
-    fn push_job(
-        env: &MockRunEnv,
-        run_id: RunId,
-        profile: &str,
-        ctx: Option<crate::types::ExecutionContext>,
-    ) {
-        env.provider.set_claim_result(run_id, ctx);
-        env.handle
-            .discover_tx
-            .send((run_id, profile.into()))
-            .unwrap();
-    }
-
-    /// Trigger graceful shutdown and wait for run() to exit.
-    async fn shutdown(env: &MockRunEnv, run_handle: tokio::task::JoinHandle<RunnerResult<()>>) {
-        env.drain();
-        env.cancel.cancel();
-        let result = tokio::time::timeout(Duration::from_secs(10), run_handle)
-            .await
-            .expect("run should finish within 10s")
-            .expect("task should not panic");
-        assert!(result.is_ok());
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 1: Normal discover → claim → execute → complete
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn main_loop_discover_claim_execute_complete() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(completion.is_some(), "job should complete");
-        let c = completion.unwrap();
-        assert_eq!(c.exit_code, 0);
-        assert!(c.error.is_none());
-
-        shutdown(&env, run_handle).await;
-    }
-
-    /// Regression for #11157: normal Running mode with available budget must
-    /// still reap completed job tasks so their cancel tokens do not remain
-    /// until a later drain, shutdown, or budget-exhausted wait.
-    #[tokio::test(start_paused = true)]
-    async fn running_reaps_completed_jobs_without_budget_exhaustion() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let run_handle = tokio::spawn(run(config));
-
-        tokio::time::timeout(
-            Duration::from_secs(2),
-            env.handle.discover_entered.notified(),
-        )
-        .await
-        .expect("run() did not enter discover_fut select! within 2s");
-
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(completion.is_some(), "job should complete");
-
-        wait_cancel_token_removed(&env.cancel_tokens, run_id, Duration::from_secs(5)).await;
-
-        shutdown(&env, run_handle).await;
-    }
-
-    /// Regression guard: the post-complete deferred network-log upload (moved
-    /// out of `post_job_cleanup` by #9828) must still reach the telemetry
-    /// endpoint, AND the drain shutdown must actually block on it — catching a
-    /// `tokio::spawn` fire-and-forget refactor that would silently lose the
-    /// upload on runtime drop.
-    ///
-    /// The mock responds with a 400 ms delay. Since the job completes almost
-    /// immediately under `MockSandboxRuntime`, `shutdown()` is invoked while
-    /// the deferred `tokio::join!(flush, upload)` is still in-flight, so a
-    /// well-behaved drain returns AFTER the mock delay elapses. A detached
-    /// upload would let shutdown return immediately — the elapsed-time
-    /// assertion below is what catches that.
-    #[tokio::test]
-    async fn deferred_network_log_upload_drains_on_graceful_shutdown() {
-        use httpmock::prelude::*;
-
-        const MOCK_DELAY: Duration = Duration::from_millis(400);
-
-        let server = MockServer::start_async().await;
-        let network_log_mock = server
-            .mock_async(|when, then| {
-                when.method(POST)
-                    .path("/api/webhooks/agent/telemetry")
-                    .body_includes("pending.example");
-                then.delay(MOCK_DELAY)
-                    .status(200)
-                    .header("content-type", "application/json")
-                    .body(r#"{"success":true,"id":"ok"}"#);
-            })
-            .await;
-
-        let (mut config, env) =
-            mock_run_config_with_api_url(test_profiles(), 8, 32768, 4, &server.base_url());
-        let write_started = Arc::new(tokio::sync::Notify::new());
-        let release_write = Arc::new(tokio::sync::Semaphore::new(0));
-        let network_log_manager =
-            NetworkLogManager::new_with_write_gate(write_started.clone(), release_write.clone());
-        Arc::get_mut(&mut config.exec_config)
-            .expect("test config should not share exec_config before run starts")
-            .network_log_manager = network_log_manager.clone();
-
-        // Seed a network log file so `upload_network_logs` has a payload to POST
-        // (otherwise it early-returns on NotFound and the assertion below would
-        // measure nothing).
-        let run_id = RunId::new_v4();
-        let network_log_path = config.exec_config.log_paths.network_log(run_id);
-        std::fs::create_dir_all(network_log_path.parent().unwrap()).unwrap();
-        std::fs::write(
-            &network_log_path,
-            concat!(
-                r#"{"timestamp":"2026-01-01T00:00:00","action":"ALLOW","host":"example.com","method":"GET","url":"https://example.com/","status":200}"#,
-                "\n",
-            ),
-        )
-        .unwrap();
-        let _network_log_session = network_log_manager
-            .register_source_ip("10.200.0.200", network_log_path.clone())
-            .await;
-        assert!(
-            network_log_manager
-                .append_for_ip(
-                    "10.200.0.200",
-                    serde_json::json!({
-                        "timestamp": "2026-01-01T00:00:01Z",
-                        "type": "dns",
-                        "host": "pending.example",
-                        "port": 53,
-                    }),
-                )
-                .await
-        );
-        write_started.notified().await;
-
-        let run_handle = tokio::spawn(run(config));
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        // The finalizer now closes Rust-side network-log attribution before
-        // completing the job, so release the accepted write before waiting for
-        // completion. The upload itself is still deferred until after the
-        // completion request below.
-        release_write.add_permits(1);
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(completion.is_some(), "job should complete");
-
-        // Drain shutdown — must block on each `spawn_job` closure's deferred
-        // `tokio::join!(flush, upload)` via the outer `jobs` JoinSet.
-        let shutdown_start = tokio::time::Instant::now();
-        shutdown(&env, run_handle).await;
-        let shutdown_elapsed = shutdown_start.elapsed();
-
-        network_log_mock.assert_calls_async(1).await;
-
-        // Stronger invariant: drain must actually WAIT for the deferred work.
-        // With a 400 ms mock delay, a well-behaved drain takes ≥ the delay;
-        // a detached (fire-and-forget) upload would let shutdown return in
-        // tens of ms, dropping the in-flight request on runtime teardown.
-        assert!(
-            shutdown_elapsed >= MOCK_DELAY - Duration::from_millis(50),
-            "drain must block on deferred upload (≥{MOCK_DELAY:?}); took only {shutdown_elapsed:?}",
-        );
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 2: Discover survives heartbeat ticks (regression #8783)
-    //
-    // ApiProvider's discover() has an internal poll timer (30s) that must
-    // survive heartbeat ticks (10s). Without pinning, `select!` cancels
-    // and recreates discover() each tick, restarting the timer from scratch.
-    //
-    // We use poll_delay=20s to simulate this: if the future is pinned, the
-    // delay completes at t=20s and the job is discovered. If not pinned,
-    // heartbeat at t=10s restarts the delay → it won't complete until t=30s.
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn discover_survives_heartbeat_ticks() {
-        let (config, env) = mock_run_config_with_delay(
-            test_profiles(),
-            8,
-            32768,
-            4,
-            Duration::from_secs(20), // poll delay: 20s
-        );
-        let run_handle = tokio::spawn(run(config));
-
-        // Push job immediately — it's in the channel, waiting for
-        // discover to finish its poll delay and read it.
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        // Advance past the 20s poll delay. Heartbeat fires at 10s but
-        // must NOT restart the delay (because discover_fut is pinned).
-        tokio::time::sleep(Duration::from_secs(25)).await;
-
-        // Heartbeat should have fired during the wait.
-        assert!(
-            env.handle.heartbeat_count() > 0,
-            "heartbeat should fire while discover poll delay is running"
-        );
-
-        // Job should have been discovered and completed.
-        // If discover was cancelled and recreated at t=10s, the 20s delay
-        // restarts → at t=25s only 15s of the second delay has elapsed →
-        // job not discovered yet → this assertion fails.
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(10))
-            .await;
-        assert!(
-            completion.is_some(),
-            "job should complete — discover must survive heartbeat ticks (regression #8783)"
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 3: Shutdown completes without deadlock (regression #8898)
-    //
-    // Uses REAL time (not paused) because a Mutex deadlock blocks the
-    // tokio runtime — paused time can't advance past a non-timer await.
-    //
-    // Only sends Draining (does NOT cancel the token). This forces the
-    // worst-case race: mode_rx.changed() wins the select!, loop breaks
-    // at the top-of-loop check, and discover_fut is never polled again.
-    // The explicit `drop(discover_fut)` releases the Mutex so shutdown()
-    // can proceed. Without that drop, shutdown() deadlocks on the Mutex.
-    // -----------------------------------------------------------------------
-
-    #[tokio::test]
-    async fn shutdown_completes_without_deadlock() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let run_handle = tokio::spawn(run(config));
-
-        // Let the main loop start and enter select!.
-        tokio::time::sleep(Duration::from_millis(50)).await;
-
-        // Only send Draining — do NOT cancel. The Draining path sees
-        // `jobs.is_empty()` immediately (no active jobs), breaks to
-        // teardown, and `drop(discover_fut)` releases the Mutex before
-        // `provider.shutdown()`. Without that drop → deadlock (regression #8898).
-        env.drain();
-
-        match tokio::time::timeout(Duration::from_secs(2), run_handle).await {
-            Ok(Ok(Ok(()))) => {}
-            Ok(Ok(Err(e))) => panic!("run() returned error: {e}"),
-            Ok(Err(e)) => panic!("task panicked: {e}"),
-            Err(_) => {
-                panic!("deadlock detected: run() did not finish within 2s (regression #8898)")
-            }
-        }
-    }
-
-    // -----------------------------------------------------------------------
-    // Draining / resume / hard-shutdown state machine
-    // -----------------------------------------------------------------------
-
-    /// SIGUSR1 → SIGUSR2 round-trip. While draining, the runner keeps the
-    /// in-flight job alive and, on resume, returns to claiming new jobs.
-    #[tokio::test]
-    async fn drain_then_resume_keeps_jobs_running() {
-        let gate = Arc::new(tokio::sync::Notify::new());
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::with_wait_exit_gate(
-            Arc::clone(&gate),
-        ));
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 32768, 4, overrides);
-        let run_handle = tokio::spawn(run(config));
-
-        // Claim a job and let it reach the gated wait.
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-        let _token = wait_cancel_token(&env.cancel_tokens, run_id, Duration::from_secs(5)).await;
-
-        // Enter Draining. The job keeps running; no cancellation is fired.
-        env.drain();
-        tokio::time::sleep(Duration::from_millis(100)).await;
-
-        // Resume. Job is still alive in the executor.
-        env.resume();
-        tokio::time::sleep(Duration::from_millis(100)).await;
-
-        // Release the gated job so it completes normally.
-        gate.notify_one();
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(completion.is_some(), "job should complete after resume");
-        let c = completion.unwrap();
-        assert_eq!(c.exit_code, 0, "job ran to normal completion");
-        assert!(c.error.is_none(), "no cancellation error");
-
-        // Runner is back in Running — a second job is claimed (cancel_token
-        // inserted). Don't wait for completion here; the shared wait_exit_gate
-        // would also block this job's exit.
-        let run_id_2 = RunId::new_v4();
-        push_job(
-            &env,
-            run_id_2,
-            "vm0/default",
-            Some(minimal_context(run_id_2)),
-        );
-        let _token_2 =
-            wait_cancel_token(&env.cancel_tokens, run_id_2, Duration::from_secs(5)).await;
-
-        // Tear down hard — the shared gate would otherwise block the
-        // second job's natural completion during Draining.
-        env.trigger_stopping().await;
-        let result = tokio::time::timeout(Duration::from_secs(5), run_handle)
-            .await
-            .expect("run should exit within 5s after hard shutdown")
-            .expect("task should not panic");
-        assert!(result.is_ok());
-    }
-
-    /// Regression guard for the unified reactor's Draining-entry state.
-    ///
-    /// The first SIGUSR1 drains the idle pool, then SIGUSR2 resumes Running.
-    /// A later job completion parks a VM, and the second SIGUSR1 must drain
-    /// that newly parked VM. If `draining_idle_pool_drained` is not reset on
-    /// Running, the second drain skips idle-pool cleanup and leaks budget.
-    #[tokio::test]
-    async fn drain_resume_then_second_drain_drains_idle_pool() {
-        let gate = Arc::new(tokio::sync::Notify::new());
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::with_wait_exit_gate(
-            Arc::clone(&gate),
-        ));
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 32768, 4, overrides);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let budget = Arc::clone(&config.budget);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-second-drain")),
-        );
-        let _token = wait_cancel_token(&env.cancel_tokens, run_id, Duration::from_secs(5)).await;
-
-        env.drain();
-        wait_parking_state(
-            &idle_pool,
-            ParkingState::SoftDraining,
-            Duration::from_secs(5),
-        )
-        .await;
-        assert_eq!(*env.mode_tx.borrow(), RunnerMode::Draining);
-
-        env.resume();
-        wait_parking_state(&idle_pool, ParkingState::Open, Duration::from_secs(5)).await;
-        assert_eq!(*env.mode_tx.borrow(), RunnerMode::Running);
-
-        gate.notify_one();
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(completion.is_some(), "job should complete after resume");
-        assert_eq!(idle_pool.lock().await.len(), 1, "job should park a VM");
-        assert_eq!(
-            budget.allocated().2,
-            1,
-            "parked VM should hold a budget slot"
-        );
-
-        env.drain();
-        match tokio::time::timeout(Duration::from_secs(5), run_handle).await {
-            Ok(Ok(Ok(()))) => {}
-            Ok(Ok(Err(e))) => panic!("run() returned error: {e}"),
-            Ok(Err(e)) => panic!("task panicked: {e}"),
-            Err(_) => panic!("second drain should exit within 5s"),
-        }
-
-        assert_eq!(
-            idle_pool.lock().await.len(),
-            0,
-            "second drain must clear the idle pool",
-        );
-        assert_eq!(
-            budget.allocated().2,
-            0,
-            "second drain must release the parked VM budget",
-        );
-    }
-
-    /// Invariant: heartbeat ticks must fire while the unified reactor is
-    /// parked in Draining mode. Silently dropping its `heartbeat_tick` branch
-    /// would leave a draining runner looking dead to the server until it exits.
-    ///
-    /// Drain before the first tick (t >= 10s) so the runner transitions to
-    /// Draining mode first; the tick observed after the time advance therefore
-    /// had to be handled by the Draining-mode heartbeat branch.
-    #[tokio::test(start_paused = true)]
-    async fn heartbeat_fires_while_draining() {
-        let gate = Arc::new(tokio::sync::Notify::new());
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::with_wait_exit_gate(
-            Arc::clone(&gate),
-        ));
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 32768, 4, overrides);
-        let run_handle = tokio::spawn(run(config));
-
-        // Claim a gated job so Draining mode has an active job to wait
-        // on — otherwise `jobs.is_empty()` auto-transitions straight to
-        // Stopping before the Draining wait path is exercised.
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-        let _token = wait_cancel_token(&env.cancel_tokens, run_id, Duration::from_secs(5)).await;
-
-        // Enter Draining before the first heartbeat tick fires. The sleep
-        // lets the runner observe `mode_rx.changed()` and re-enter the
-        // reactor with Draining-mode guards. There is no production-side
-        // notifier for "Draining mode entered", so this
-        // synchronization has to be time-based.
-        env.drain();
-        tokio::time::sleep(Duration::from_millis(100)).await;
-        assert_eq!(*env.mode_tx.borrow(), RunnerMode::Draining);
-        let before = env.handle.heartbeat_count();
-
-        // Advance past the first tick while Draining mode is active.
-        // A broken Draining path that dropped its `heartbeat_tick.tick()`
-        // branch would leave the count unchanged; `wait_heartbeat_past`
-        // returns false on timeout.
-        tokio::time::advance(HEARTBEAT_PERIOD + Duration::from_secs(5)).await;
-        assert!(
-            env.handle
-                .wait_heartbeat_past(before, Duration::from_secs(5))
-                .await,
-            "Draining mode must handle heartbeat_tick (baseline={before})",
-        );
-
-        // Tear down hard — the gate would block natural completion.
-        env.trigger_stopping().await;
-        let _ = tokio::time::timeout(Duration::from_secs(5), run_handle).await;
-    }
-
-    /// Invariant: heartbeat ticks must fire while the unified reactor is
-    /// parked with budget exhausted. Dropping its `heartbeat_tick` branch would
-    /// make a runner that's at resource capacity look dead to the server until
-    /// budget frees.
-    ///
-    /// A 1-slot budget + a gated job pins the runner in the budget-exhausted
-    /// state for the duration of the time advance.
-    #[tokio::test(start_paused = true)]
-    async fn heartbeat_fires_while_budget_exhausted() {
-        let gate = Arc::new(tokio::sync::Notify::new());
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::with_wait_exit_gate(
-            Arc::clone(&gate),
-        ));
-        // Budget sized for exactly one `test_profiles()` slot (vcpu=2, mem=4096).
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 2, 4096, 1, overrides);
-        let budget = Arc::clone(&config.budget);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-        // Wait for the reservation — after this, the next loop iteration
-        // enters the budget-exhausted wait state at the can_afford check.
-        // The sleep yields to the runner so it reaches the reactor `select!`
-        // before the time advance below.
-        wait_budget_count(&budget, 1, Duration::from_secs(5)).await;
-        tokio::time::sleep(Duration::from_millis(100)).await;
-        let before = env.handle.heartbeat_count();
-
-        // Advance past the first tick while the runner is budget-exhausted.
-        // Removing the `heartbeat_tick.tick()` branch from the reactor `select!`
-        // leaves the count unchanged; `wait_heartbeat_past` returns false
-        // on timeout.
-        tokio::time::advance(HEARTBEAT_PERIOD + Duration::from_secs(5)).await;
-        assert!(
-            env.handle
-                .wait_heartbeat_past(before, Duration::from_secs(5))
-                .await,
-            "budget-exhausted arm must handle heartbeat_tick (baseline={before})",
-        );
-
-        // Release the gate so the job completes, budget frees, and the
-        // standard `shutdown()` helper (Draining → auto-Stop) terminates
-        // the runner cleanly — same pattern as `budget_full_skips_then_resumes`.
-        gate.notify_one();
-        let _ = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        shutdown(&env, run_handle).await;
-    }
-
-    /// With no active jobs, SIGUSR1 transitions straight through Draining
-    /// and exits within a few hundred ms.
-    #[tokio::test]
-    async fn drain_without_active_jobs_exits_promptly() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let run_handle = tokio::spawn(run(config));
-
-        tokio::time::sleep(Duration::from_millis(50)).await;
-        env.drain();
-
-        match tokio::time::timeout(Duration::from_secs(2), run_handle).await {
-            Ok(Ok(Ok(()))) => {}
-            Ok(Ok(Err(e))) => panic!("run() returned error: {e}"),
-            Ok(Err(e)) => panic!("task panicked: {e}"),
-            Err(_) => panic!("drain with no active jobs should exit within 2s"),
-        }
-    }
-
-    /// SIGUSR2 on an already-Running runner is a no-op: it does not disturb
-    /// normal discovery.
-    #[tokio::test(start_paused = true)]
-    async fn resume_on_running_is_noop() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let run_handle = tokio::spawn(run(config));
-
-        // SIGUSR2 while already Running — state guard blocks the send,
-        // leaving mode unchanged and discovery uninterrupted.
-        env.resume();
-        tokio::time::sleep(Duration::from_millis(50)).await;
-
-        // Runner is still claiming jobs.
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(
-            completion.is_some(),
-            "resume on Running should not break discovery"
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    /// SIGTERM while a job is in flight: per-job cancellation fires, the
-    /// executor aborts, and run() exits within a couple of seconds rather
-    /// than blocking on the 2h JOB_TIMEOUT.
-    #[tokio::test]
-    async fn hard_shutdown_cancels_active_jobs() {
-        let gate = Arc::new(tokio::sync::Notify::new());
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::with_wait_exit_gate(
-            gate,
-        ));
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 32768, 4, overrides);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        // Wait for the job to enter the gated wait — cancel token is now in the map.
-        let _token = wait_cancel_token(&env.cancel_tokens, run_id, Duration::from_secs(5)).await;
-
-        // SIGTERM equivalent: latch hard-shutdown, cancel all in-flight jobs.
-        env.trigger_stopping().await;
-
-        match tokio::time::timeout(Duration::from_secs(3), run_handle).await {
-            Ok(Ok(Ok(()))) => {}
-            Ok(Ok(Err(e))) => panic!("run() returned error: {e}"),
-            Ok(Err(e)) => panic!("task panicked: {e}"),
-            Err(_) => panic!("hard shutdown should exit within 3s — got stuck"),
-        }
-
-        // The cancelled job reports the synthetic "cancelled by user" error.
-        let comps = env.handle.completions.lock().unwrap();
-        let c = comps
-            .iter()
-            .find(|c| c.run_id == run_id)
-            .expect("cancelled job should still report completion");
-        assert_eq!(c.error.as_deref(), Some("cancelled by user"));
-    }
-
-    /// SIGUSR1 → SIGTERM upgrade. Starts Draining, then hard-shutdown fires
-    /// mid-drain and the run exits promptly with the active job cancelled.
-    #[tokio::test]
-    async fn drain_then_hard_shutdown_upgrades() {
-        let gate = Arc::new(tokio::sync::Notify::new());
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::with_wait_exit_gate(
-            gate,
-        ));
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 32768, 4, overrides);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-        let _token = wait_cancel_token(&env.cancel_tokens, run_id, Duration::from_secs(5)).await;
-
-        // Draining. Without hard shutdown, this would wait up to JOB_TIMEOUT = 2h.
-        env.drain();
-        tokio::time::sleep(Duration::from_millis(100)).await;
-
-        // Upgrade to hard shutdown.
-        env.trigger_stopping().await;
-
-        match tokio::time::timeout(Duration::from_secs(3), run_handle).await {
-            Ok(Ok(Ok(()))) => {}
-            Ok(Ok(Err(e))) => panic!("run() returned error: {e}"),
-            Ok(Err(e)) => panic!("task panicked: {e}"),
-            Err(_) => panic!("Draining → hard shutdown should exit within 3s"),
-        }
-    }
-
-    /// TOCTOU regression: a SIGTERM that iterates `cancel_tokens` *before*
-    /// the main loop inserts a newly-claimed job's token would leave that
-    /// job running uncancelled. The fix is a post-insert `mode_rx.borrow()`
-    /// check that catches Stopping and cancels the token in that window.
-    ///
-    /// To reproduce deterministically, we use `send_if_modified` to flip
-    /// the watch value to `Stopping` **without** waking `mode_rx.changed()`
-    /// — this is exactly what the racy window looks like to the main loop:
-    /// its outer select! is still polling discover_fut, unaware that the
-    /// value has changed. When discover yields a job, the main loop takes
-    /// the claim path, inserts the token, then reads `mode_rx.borrow()`
-    /// and catches the Stopping value that was silently written.
-    #[tokio::test]
-    async fn claim_after_stopping_sent_cancels_new_job() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let run_handle = tokio::spawn(run(config));
-
-        // Deterministic barrier: wait for run()'s main loop to have polled
-        // `discover_fut` into its await state. Only then is the Running-mode
-        // reactor `select!` provably in place, which is the precondition for the
-        // silent `send_if_modified` below to land without waking the loop.
-        // A wall-clock sleep here flakes under coverage CI — see #10146.
-        // The 2s timeout gives a clear diagnostic if the "loop parks on
-        // discover" invariant ever regresses, rather than hanging until
-        // the outer test harness kills us.
-        tokio::time::timeout(
-            Duration::from_secs(2),
-            env.handle.discover_entered.notified(),
-        )
-        .await
-        .expect("run() did not enter discover_fut select! within 2s");
-
-        // Flip the watch value to Stopping without firing changed().
-        env.parking_gate.close();
-        env.mode_tx.send_if_modified(|v| {
-            *v = RunnerMode::Stopping;
-            false
-        });
-
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        // `wait_completion` is event-driven (fires on `provider.complete`), so
-        // this duration is a diagnostic cap for genuine hangs — not a budget
-        // for the run. A large cap absorbs coverage-CI slowdown of the full
-        // dispatch→executor→complete chain without flaking (see #10146).
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(30))
-            .await;
-        assert!(
-            c.is_some(),
-            "job must report cancellation even when the handler missed the token"
-        );
-        assert_eq!(c.unwrap().error.as_deref(), Some("cancelled by user"));
-
-        // Let run() exit — fire changed() now so the main loop observes
-        // Stopping at loop top and breaks to teardown.
-        env.parking_gate.close();
-        env.mode_tx.send_modify(|v| {
-            *v = RunnerMode::Stopping;
-        });
-        env.cancel.cancel();
-        let _ = tokio::time::timeout(Duration::from_secs(5), run_handle).await;
-    }
-
-    /// SIGUSR2 received while Stopping is committed is ignored — the
-    /// runner cannot resume out of Stopping.
-    #[tokio::test]
-    async fn resume_after_stopping_is_ignored() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let run_handle = tokio::spawn(run(config));
-
-        // Wait for the main loop to park on `discover_fut` so the subsequent
-        // `trigger_stopping` lands on a steady-state loop rather than racing
-        // against startup. This test does not depend on the silent-flip
-        // semantics of `claim_after_stopping_sent_cancels_new_job` (it uses
-        // `trigger_stopping`, which fires `changed()`), but the same barrier
-        // is still the right "main loop is idle" signal — and deterministic
-        // under coverage CI, unlike the 50 ms sleep this replaces.
-        tokio::time::timeout(
-            Duration::from_secs(2),
-            env.handle.discover_entered.notified(),
-        )
-        .await
-        .expect("run() did not enter discover_fut select! within 2s");
-
-        // Enter Stopping first.
-        env.trigger_stopping().await;
-
-        // handle_resume_signal refuses any transition except from Draining.
-        handle_resume_signal(&env.lifecycle);
-        assert_eq!(
-            *env.mode_tx.borrow(),
-            RunnerMode::Stopping,
-            "mode must remain Stopping after ignored SIGUSR2"
-        );
-
-        match tokio::time::timeout(Duration::from_secs(2), run_handle).await {
-            Ok(Ok(Ok(()))) => {}
-            Ok(Ok(Err(e))) => panic!("run() returned error: {e}"),
-            Ok(Err(e)) => panic!("task panicked: {e}"),
-            Err(_) => panic!("hard shutdown should exit within 2s"),
-        }
-    }
-
-    /// Regression for #10146 / #10223: the main-loop `idle_cleanup` and
-    /// `heartbeat_tick` intervals must defer their first tick past the
-    /// configured period, so neither tick branch is Ready on the first `select!`
-    /// poll. Otherwise they pre-empt `discover_fut` (which parks on
-    /// `rx.recv()` → Pending) and any silent `mode_tx` flip during the
-    /// tick body breaks the loop before the pending job is ever claimed.
-    ///
-    /// The behavioral test `claim_after_stopping_sent_cancels_new_job`
-    /// only triggers the underlying race under `cargo llvm-cov`, so a
-    /// silent revert of `interval_at` → `interval` would not fail it on
-    /// the default CI path. This test pins the invariant directly: a
-    /// job pushed immediately at startup is processed without any tick
-    /// having fired, observable via `heartbeat_count == 0`.
-    #[tokio::test(start_paused = true)]
-    async fn heartbeat_tick_defers_past_first_select_poll() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let run_handle = tokio::spawn(run(config));
-
-        tokio::time::timeout(
-            Duration::from_secs(2),
-            env.handle.discover_entered.notified(),
-        )
-        .await
-        .expect("run() did not enter discover_fut select! within 2s");
-
-        // `minimal_context` → no session → completion path does not trigger
-        // `park_notify`, so any heartbeat observed here came from the
-        // interval tick (the path we want to prove did NOT fire).
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(completion.is_some(), "job must complete");
-
-        assert_eq!(
-            env.handle.heartbeat_count(),
-            0,
-            "heartbeat tick fired before the startup job was processed — \
-             is the main-loop interval `interval_at(now + period, period)` \
-             instead of `interval(period)`?"
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    /// Draining auto-transitions to Stopping when jobs drain naturally.
-    /// Verifies the internal lifecycle transition from Draining to Stopping.
-    #[tokio::test]
-    async fn drain_with_jobs_transitions_to_stopping_when_empty() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let run_handle = tokio::spawn(run(config));
-
-        // Let a quick job complete, then drain.
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-        let _ = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-
-        // Give the main loop a moment to clean up the jobset, then drain.
-        tokio::time::sleep(Duration::from_millis(50)).await;
-        env.drain();
-
-        // Draining mode should observe jobs.is_empty() and self-send
-        // Stopping, leading to teardown and run() exit.
-        match tokio::time::timeout(Duration::from_secs(3), run_handle).await {
-            Ok(Ok(Ok(()))) => {}
-            Ok(Ok(Err(e))) => panic!("run() returned error: {e}"),
-            Ok(Err(e)) => panic!("task panicked: {e}"),
-            Err(_) => panic!("Draining natural drain should exit within 3s"),
-        }
-
-        assert_eq!(
-            *env.mode_tx.borrow(),
-            RunnerMode::Stopping,
-            "mode_tx must reflect Stopping after natural drain transition"
-        );
-        assert_eq!(env.parking_gate.state(), ParkingState::Closed);
-
-        // Observability pin: the Draining → Stopping auto-transition must
-        // emit a one-shot heartbeat with mode="stopping" before teardown,
-        // in addition to the terminal heartbeat during teardown. Two or
-        // more "stopping" heartbeats prove both sites fire (the one-shot
-        // at the transition and the terminal one). A single hit would mean
-        // one of the two was removed.
-        let stopping_count = env
-            .handle
-            .heartbeats
-            .lock()
-            .unwrap()
-            .iter()
-            .filter(|h| h.mode == "stopping")
-            .count();
-        assert!(
-            stopping_count >= 2,
-            "expected at least 2 stopping heartbeats (one-shot + terminal), got {stopping_count}",
-        );
-    }
-
-    /// Race regression: the Draining → Stopping auto-transition must be
-    /// guarded on `mode == Draining`, so a concurrent SIGUSR2 that flips
-    /// mode back to Running is preserved rather than silently overwritten.
-    ///
-    /// We simulate the race deterministically:
-    /// 1. Claim a gated job — mode is Draining and the reactor is waiting
-    ///    with Draining-mode guards.
-    /// 2. Silently flip mode to Running via `send_if_modified(false)`
-    ///    (equivalent to SIGUSR2 arriving *after* the arm noticed jobs was
-    ///    non-empty but *before* the next iteration's guard).
-    /// 3. Release the gate — the job completes, the reactor reaps it, loops to
-    ///    top, sees `jobs.is_empty()`, and evaluates the guarded
-    ///    `send_if_modified`. The guard rejects the overwrite because mode
-    ///    is no longer Draining.
-    /// 4. Outer loop re-reads mode → Running → resumes normal discovery.
-    #[tokio::test]
-    async fn draining_auto_stop_preserves_concurrent_resume() {
-        let gate = Arc::new(tokio::sync::Notify::new());
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::with_wait_exit_gate(
-            Arc::clone(&gate),
-        ));
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 32768, 4, overrides);
-        let run_handle = tokio::spawn(run(config));
-
-        // Claim a job and hold it at the gate so Draining mode has
-        // something to wait on — without a live job the auto-transition
-        // fires before any concurrent signal could race.
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-        let _token = wait_cancel_token(&env.cancel_tokens, run_id, Duration::from_secs(5)).await;
-
-        env.drain();
-        tokio::time::sleep(Duration::from_millis(50)).await;
-        assert_eq!(*env.mode_tx.borrow(), RunnerMode::Draining);
-
-        // Silently flip to Running — the `false` return suppresses
-        // `changed()`, so the arm does not wake on a mode transition. The
-        // guard will only observe the new value on its next iteration's
-        // send_if_modified closure.
-        env.parking_gate.open_after_soft_drain();
-        env.mode_tx.send_if_modified(|v| {
-            *v = RunnerMode::Running;
-            false
-        });
-
-        // Release the gate: job completes, the arm reaps, then checks
-        // jobs.is_empty() → true → calls the guarded send_if_modified.
-        gate.notify_one();
-        let _ = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        tokio::time::sleep(Duration::from_millis(100)).await;
-
-        assert_eq!(
-            *env.mode_tx.borrow(),
-            RunnerMode::Running,
-            "SIGUSR2 must win the race against the Draining auto-Stop",
-        );
-
-        // Tear down cleanly.
-        env.trigger_stopping().await;
-        let _ = tokio::time::timeout(Duration::from_secs(5), run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 4: Claim failure (409) rolls back budget
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn claim_failure_rolls_back_budget() {
-        // Budget for exactly 1 job (2 vcpu, 4096 MB matches the test profile).
-        let (config, env) = mock_run_config(test_profiles(), 2, 4096, 1);
-        let run_handle = tokio::spawn(run(config));
-
-        // First job: claim returns None (409 conflict)
-        let run_id_1 = RunId::new_v4();
-        push_job(&env, run_id_1, "vm0/default", None);
-
-        // Give main loop time to process the failed claim and release budget.
-        tokio::time::advance(Duration::from_millis(100)).await;
-        tokio::task::yield_now().await;
-
-        // Second job: claim succeeds — budget should have been freed.
-        let run_id_2 = RunId::new_v4();
-        push_job(
-            &env,
-            run_id_2,
-            "vm0/default",
-            Some(minimal_context(run_id_2)),
-        );
-
-        let completion = env
-            .handle
-            .wait_completion(run_id_2, Duration::from_secs(5))
-            .await;
-        assert!(
-            completion.is_some(),
-            "second job should complete (budget freed after first 409)"
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 5: Shutdown drains running jobs before exiting
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn shutdown_drains_running_jobs() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        // Wait for completion before draining.
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(completion.is_some());
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 6: Unknown profile is skipped without affecting subsequent jobs
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn unknown_profile_skipped() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let run_handle = tokio::spawn(run(config));
-
-        // Push a job with a profile that doesn't exist in the profiles map.
-        // The main loop should log a warning and continue without claiming.
-        let bad_id = RunId::new_v4();
-        push_job(
-            &env,
-            bad_id,
-            "vm0/nonexistent",
-            Some(minimal_context(bad_id)),
-        );
-
-        // Give main loop time to skip the bad job.
-        tokio::time::sleep(Duration::from_millis(100)).await;
-
-        // Push a valid job — it should succeed despite the earlier bad one.
-        let good_id = RunId::new_v4();
-        push_job(&env, good_id, "vm0/default", Some(minimal_context(good_id)));
-
-        let completion = env
-            .handle
-            .wait_completion(good_id, Duration::from_secs(5))
-            .await;
-        assert!(
-            completion.is_some(),
-            "valid job should complete after unknown profile is skipped"
-        );
-
-        // The bad job should never have been claimed (no completion recorded).
-        {
-            let comps = env.handle.completions.lock().unwrap();
-            assert!(
-                !comps.iter().any(|c| c.run_id == bad_id),
-                "unknown-profile job should not produce a completion"
-            );
-        }
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 7: Duplicate discovery (same run_id) is deduplicated
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn duplicate_discovery_deduplicated() {
-        // Budget for 2 jobs — enough for the duplicate to pass the budget
-        // check and reach the cancel_tokens dedup logic.
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let budget = Arc::clone(&config.budget);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        // Wait for job to be claimed and executing (cancel_tokens now has run_id).
-        tokio::time::sleep(Duration::from_millis(100)).await;
-
-        // Push the same run_id again (simulates duplicate discovery).
-        // Budget has room, but cancel_tokens already contains this run_id →
-        // the duplicate is rejected and budget is released.
-        env.handle
-            .discover_tx
-            .send((run_id, "vm0/default".into()))
-            .unwrap();
-
-        // Wait for the original job to complete.
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(completion.is_some(), "original job should complete");
-
-        // Only one completion should exist for this run_id.
-        {
-            let comps = env.handle.completions.lock().unwrap();
-            let count = comps.iter().filter(|c| c.run_id == run_id).count();
-            assert_eq!(
-                count, 1,
-                "duplicate discovery should not produce a second completion"
-            );
-        }
-        wait_budget_count(&budget, 0, Duration::from_secs(5)).await;
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 8: Two successful jobs in sequence
-    //
-    // After the first job completes, discover_fut is recreated
-    // (Box::pin(provider.discover())). The second job must be discovered,
-    // claimed, executed, and completed through the recreated future.
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn two_sequential_jobs_complete() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let run_handle = tokio::spawn(run(config));
-
-        // First job
-        let id1 = RunId::new_v4();
-        push_job(&env, id1, "vm0/default", Some(minimal_context(id1)));
-        let c1 = env
-            .handle
-            .wait_completion(id1, Duration::from_secs(5))
-            .await;
-        assert!(c1.is_some(), "first job should complete");
-        assert_eq!(c1.unwrap().exit_code, 0);
-
-        // Second job — exercises the recreated discover_fut path
-        let id2 = RunId::new_v4();
-        push_job(&env, id2, "vm0/default", Some(minimal_context(id2)));
-        let c2 = env
-            .handle
-            .wait_completion(id2, Duration::from_secs(5))
-            .await;
-        assert!(
-            c2.is_some(),
-            "second job should complete via recreated discover_fut"
-        );
-        assert_eq!(c2.unwrap().exit_code, 0);
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 9: idle pool park/take is gated on session ID availability
-    //
-    // With a session ID, the VM is parked after execution; without one,
-    // the VM is destroyed (no key to re-find it under).
-    // -----------------------------------------------------------------------
-
-    fn context_with_session_opt(
-        run_id: RunId,
-        session_id: Option<&str>,
-    ) -> crate::types::ExecutionContext {
-        let mut ctx = minimal_context(run_id);
-        if let Some(sid) = session_id {
-            ctx.resume_session = Some(crate::types::ResumeSession {
-                session_id: sid.to_string(),
-                session_history: String::new(),
-            });
-        }
-        ctx
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn job_with_session_parks_vm() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        let ctx = context_with_session_opt(run_id, Some("sess-1"));
-        push_job(&env, run_id, "vm0/default", Some(ctx));
-
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(c.is_some(), "job should complete");
-        assert_eq!(c.unwrap().exit_code, 0);
-
-        let pool = env.idle_pool.lock().await;
-        assert_eq!(pool.len(), 1, "VM should be parked when session is present");
-        assert!(pool.held_sessions().contains(&"sess-1".to_string()));
-        drop(pool);
-
-        shutdown(&env, run_handle).await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn job_without_session_does_not_park() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        // No session — parking requires a session ID.
-        let ctx = context_with_session_opt(run_id, None);
-        push_job(&env, run_id, "vm0/default", Some(ctx));
-
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(c.is_some(), "job should complete");
-        assert_eq!(c.unwrap().exit_code, 0);
-
-        let pool = env.idle_pool.lock().await;
-        assert_eq!(
-            pool.len(),
-            0,
-            "VM should NOT be parked without a session ID"
-        );
-        drop(pool);
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 11: Budget full → job skipped (not claimed) → budget freed → next job succeeds
-    //
-    // Different from test 4 (claim failure): here try_reserve returns false
-    // so claim() is never called. The job stays in the channel but the main
-    // loop moves on. After the running job completes and frees budget, the
-    // next discover picks up the waiting job.
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn budget_full_skips_then_resumes() {
-        // Budget for exactly 1 job (2 vcpu, 4096 MB).
-        let (config, env) = mock_run_config(test_profiles(), 2, 4096, 1);
-        let run_handle = tokio::spawn(run(config));
-
-        // First job: claims the entire budget.
-        let id1 = RunId::new_v4();
-        push_job(&env, id1, "vm0/default", Some(minimal_context(id1)));
-
-        // Wait for job 1 to be claimed (budget now full).
-        tokio::time::sleep(Duration::from_millis(100)).await;
-
-        // Second job: pushed while budget is full. try_reserve fails →
-        // the job is skipped without claim. But it remains in the channel.
-        let id2 = RunId::new_v4();
-        push_job(&env, id2, "vm0/default", Some(minimal_context(id2)));
-
-        // Job 1 completes (MockSandbox is instant) → budget freed.
-        let c1 = env
-            .handle
-            .wait_completion(id1, Duration::from_secs(5))
-            .await;
-        assert!(c1.is_some(), "first job should complete");
-
-        // After budget is freed, the main loop re-enters the normal select!
-        // and discovers job 2 from the channel.
-        let c2 = env
-            .handle
-            .wait_completion(id2, Duration::from_secs(5))
-            .await;
-        assert!(
-            c2.is_some(),
-            "second job should complete after budget is freed"
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    /// Budget-exhausted mode must not poll discovery. A queued job should
-    /// remain undiscovered until a running job frees budget, otherwise the
-    /// runner may claim work it cannot admit.
-    #[tokio::test(start_paused = true)]
-    async fn budget_exhausted_buffers_discovery_until_budget_frees() {
-        let gate = Arc::new(tokio::sync::Notify::new());
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::with_wait_exit_gate(
-            Arc::clone(&gate),
-        ));
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 2, 4096, 1, overrides);
-        let budget = Arc::clone(&config.budget);
-        let run_handle = tokio::spawn(run(config));
-        env.handle.discover_entered.notified().await;
-
-        let id1 = RunId::new_v4();
-        push_job(&env, id1, "vm0/default", Some(minimal_context(id1)));
-        let _token_1 = wait_cancel_token(&env.cancel_tokens, id1, Duration::from_secs(5)).await;
-        wait_budget_count(&budget, 1, Duration::from_secs(5)).await;
-
-        let id2 = RunId::new_v4();
-        push_job(&env, id2, "vm0/default", Some(minimal_context(id2)));
-        tokio::time::timeout(
-            Duration::from_millis(100),
-            env.handle.discover_entered.notified(),
-        )
-        .await
-        .expect_err("discovery must not be polled while budget is exhausted");
-        assert!(
-            !env.cancel_tokens.lock().await.contains_key(&id2),
-            "queued job must not be claimed while budget is exhausted",
-        );
-        assert!(
-            env.handle
-                .completions
-                .lock()
-                .unwrap()
-                .iter()
-                .all(|c| c.run_id != id2),
-            "queued job must not complete before budget frees",
-        );
-
-        gate.notify_one();
-        let c1 = env
-            .handle
-            .wait_completion(id1, Duration::from_secs(5))
-            .await;
-        assert!(c1.is_some(), "first job should complete");
-
-        let _token_2 = wait_cancel_token(&env.cancel_tokens, id2, Duration::from_secs(5)).await;
-        gate.notify_one();
-        let c2 = env
-            .handle
-            .wait_completion(id2, Duration::from_secs(5))
-            .await;
-        assert!(
-            c2.is_some(),
-            "queued job should complete after budget is freed",
-        );
-        wait_budget_count(&budget, 0, Duration::from_secs(5)).await;
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // =======================================================================
-    // Subsystem integration tests — Phase 3
-    //
-    // Tests 10-19 cover job lifecycle (park/destroy), idle pool integration
-    // (session affinity, profile mismatch, expiry), budget exhaustion
-    // (eviction), shutdown drain, and edge cases (pool-full, reuse cycle).
-    // =======================================================================
-
-    /// ExecutionContext with a resume_session for idle pool testing.
-    fn context_with_session(run_id: RunId, session_id: &str) -> crate::types::ExecutionContext {
-        let mut ctx = minimal_context(run_id);
-        ctx.resume_session = Some(crate::types::ResumeSession {
-            session_id: session_id.into(),
-            session_history: String::new(),
-        });
-        ctx
-    }
-
-    /// Two profiles with different resource requirements for mismatch tests.
-    fn two_profiles() -> BTreeMap<String, config::ProfileConfig> {
-        let mut m = BTreeMap::new();
-        m.insert(
-            "vm0/default".to_string(),
-            config::ProfileConfig {
-                rootfs_hash: "hash".into(),
-                snapshot_hash: "snap".into(),
-                vcpu: 2,
-                memory_mb: 4096,
-                disk_mb: 10240,
-            },
-        );
-        m.insert(
-            "vm0/large".to_string(),
-            config::ProfileConfig {
-                rootfs_hash: "hash2".into(),
-                snapshot_hash: "snap2".into(),
-                vcpu: 4,
-                memory_mb: 8192,
-                disk_mb: 20480,
-            },
-        );
-        m
-    }
-
-    /// Build an active-owned park candidate with all seed fields configurable.
-    fn make_test_park_candidate(
-        session_id: &str,
-        profile_name: &str,
-        budget_lease: BudgetLease,
-    ) -> ParkCandidate {
-        ParkCandidate::from_parked_parts(ParkCandidateParts {
-            sandbox: Box::new(MockSandbox::new("idle-test")),
-            factory: Arc::new(Box::new(MockSandboxFactory::new()) as Box<dyn SandboxFactory>),
-            session_id: session_id.into(),
-            sandbox_id: SandboxId::new_v4(),
-            profile_name: profile_name.into(),
-            budget_lease,
-            source_ip: "10.0.0.1".into(),
-            storage_fingerprints: crate::idle_pool::StorageFingerprints::default(),
-        })
-    }
-
-    /// Pre-populate idle pool with an entry and reserve its budget. Returns
-    /// the entry's sandbox id so reuse tests can assert it propagates through
-    /// to the completion payload.
-    async fn seed_idle_pool(
-        pool: &SharedIdlePool,
-        budget: &Arc<ResourceBudget>,
-        session_id: &str,
-        profile_name: &str,
-        vcpu: u32,
-        memory_mb: u32,
-    ) -> SandboxId {
-        let budget_lease = ResourceBudget::try_reserve_lease(budget, vcpu, memory_mb).unwrap();
-        let candidate = make_test_park_candidate(session_id, profile_name, budget_lease);
-        let sandbox_id = candidate.sandbox_id();
-        let mut guard = pool.lock().await;
-        let result = guard.park(candidate);
-        assert!(matches!(result, ParkResult::Parked));
-        sandbox_id
-    }
-
-    async fn seed_idle_pool_with_overrides(
-        pool: &SharedIdlePool,
-        budget: &Arc<ResourceBudget>,
-        overrides: &Arc<sandbox_mock::MockSandboxOverrides>,
-        session_id: &str,
-        profile_name: &str,
-        vcpu: u32,
-        memory_mb: u32,
-    ) -> SandboxId {
-        let runtime = sandbox_mock::MockSandboxRuntime::with_overrides(Arc::clone(overrides));
-        let mut factory = runtime
-            .create_factory(sandbox::FactoryConfig {
-                profile: profile_name.into(),
-                binary_path: PathBuf::new(),
-                kernel_path: PathBuf::new(),
-                rootfs_path: PathBuf::new(),
-                base_dir: PathBuf::new(),
-                snapshot: None,
-            })
-            .await
-            .expect("create factory");
-        factory.startup().await.expect("startup");
-        let factory_arc: Arc<Box<dyn sandbox::SandboxFactory>> = Arc::new(factory);
-        let sandbox_id = SandboxId::new_v4();
-        let sandbox = factory_arc
-            .create(sandbox::SandboxConfig {
-                id: sandbox_id,
-                resources: sandbox::ResourceLimits {
-                    cpu_count: vcpu,
-                    memory_mb,
-                },
-            })
-            .await
-            .expect("create sandbox");
-        let budget_lease =
-            ResourceBudget::try_reserve_lease(budget, vcpu, memory_mb).expect("reserve budget");
-
-        let mut guard = pool.lock().await;
-        let result = guard.park(ParkCandidate::from_parked_parts(ParkCandidateParts {
-            sandbox,
-            factory: factory_arc,
-            session_id: session_id.to_string(),
-            sandbox_id,
-            profile_name: profile_name.into(),
-            budget_lease,
-            source_ip: "10.0.0.1".into(),
-            storage_fingerprints: crate::idle_pool::StorageFingerprints::default(),
-        }));
-        assert!(matches!(result, ParkResult::Parked));
-        sandbox_id
-    }
-
-    /// Poll until `budget.allocated().2` (running_count) reaches `expected`.
-    ///
-    /// The active budget lease is dropped after `provider.complete()` in the
-    /// spawned job task, so `wait_completion()` returning does NOT guarantee
-    /// the budget has been released yet. This helper avoids fixed sleeps as
-    /// synchronization.
-    async fn wait_budget_count(budget: &ResourceBudget, expected: usize, timeout: Duration) {
-        let deadline = tokio::time::Instant::now() + timeout;
-        loop {
-            if budget.allocated().2 == expected {
-                return;
-            }
-            assert!(
-                tokio::time::Instant::now() < deadline,
-                "budget count did not reach {expected} within {timeout:?} (actual: {})",
-                budget.allocated().2,
-            );
-            tokio::time::sleep(Duration::from_millis(10)).await;
-        }
-    }
-
-    async fn wait_idle_pool_len(pool: &SharedIdlePool, expected: usize, timeout: Duration) {
-        let deadline = tokio::time::Instant::now() + timeout;
-        loop {
-            let actual = pool.lock().await.len();
-            if actual == expected {
-                return;
-            }
-            assert!(
-                tokio::time::Instant::now() < deadline,
-                "idle pool length did not reach {expected} within {timeout:?} (actual: {actual})",
-            );
-            tokio::time::sleep(Duration::from_millis(10)).await;
-        }
-    }
-
-    /// Poll until the idle pool parking state reaches `expected`.
-    async fn wait_parking_state(pool: &SharedIdlePool, expected: ParkingState, timeout: Duration) {
-        let deadline = tokio::time::Instant::now() + timeout;
-        loop {
-            let actual = pool.lock().await.parking_state();
-            if actual == expected {
-                return;
-            }
-            assert!(
-                tokio::time::Instant::now() < deadline,
-                "idle pool parking state did not reach {expected:?} within {timeout:?} (actual: {actual:?})",
-            );
-            tokio::time::sleep(Duration::from_millis(10)).await;
-        }
-    }
-
-    /// Pre-populate idle pool with an expired entry (parked 400s ago, timeout 300s).
-    async fn seed_idle_pool_expired(
-        pool: &SharedIdlePool,
-        budget: &Arc<ResourceBudget>,
-        session_id: &str,
-        profile_name: &str,
-        vcpu: u32,
-        memory_mb: u32,
-    ) {
-        let budget_lease = ResourceBudget::try_reserve_lease(budget, vcpu, memory_mb).unwrap();
-        let candidate = make_test_park_candidate(session_id, profile_name, budget_lease);
-        let mut guard = pool.lock().await;
-        let result = guard.park_at_for_test(
-            candidate,
-            std::time::Instant::now() - Duration::from_secs(400),
-            Duration::from_secs(300),
-        );
-        assert!(matches!(result, ParkResult::Parked));
-    }
-
-    struct TestParkCandidateSpec<'a> {
-        session_id: &'a str,
-        profile_name: &'a str,
-        vcpu: u32,
-        memory_mb: u32,
-        parked_at: std::time::Instant,
-        idle_timeout: Duration,
-    }
-
-    async fn seed_idle_pool_with_timing(
-        pool: &SharedIdlePool,
-        budget: &Arc<ResourceBudget>,
-        spec: TestParkCandidateSpec<'_>,
-    ) {
-        let budget_lease =
-            ResourceBudget::try_reserve_lease(budget, spec.vcpu, spec.memory_mb).unwrap();
-        let candidate = make_test_park_candidate(spec.session_id, spec.profile_name, budget_lease);
-        let mut guard = pool.lock().await;
-        let result = guard.park_at_for_test(candidate, spec.parked_at, spec.idle_timeout);
-        assert!(matches!(result, ParkResult::Parked));
-    }
-
-    async fn status_idle_sessions_and_active_runs(
-        status_path: &std::path::Path,
-    ) -> (Vec<String>, Vec<String>) {
-        let raw = tokio::fs::read_to_string(status_path).await.unwrap();
-        let status: serde_json::Value = serde_json::from_str(&raw).unwrap();
-        let mut sessions: Vec<String> = status
-            .get("idle_vms")
-            .and_then(|v| v.as_array())
-            .map(|idle_vms| {
-                idle_vms
-                    .iter()
-                    .filter_map(|vm| {
-                        vm.get("session_id")
-                            .and_then(|session| session.as_str())
-                            .map(str::to_string)
-                    })
-                    .collect()
-            })
-            .unwrap_or_default();
-        sessions.sort_unstable();
-        let mut run_ids: Vec<String> = status["active_runs"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .filter_map(|run| {
-                run.get("run_id")
-                    .and_then(|run_id| run_id.as_str())
-                    .map(str::to_string)
-            })
-            .collect();
-        run_ids.sort_unstable();
-        (sessions, run_ids)
-    }
-
-    async fn status_active_run_records(status_path: &std::path::Path) -> Vec<(String, String)> {
-        let raw = tokio::fs::read_to_string(status_path).await.unwrap();
-        let status: serde_json::Value = serde_json::from_str(&raw).unwrap();
-        let mut records: Vec<(String, String)> = status["active_runs"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .map(|run| {
-                (
-                    run["run_id"].as_str().unwrap().to_string(),
-                    run["sandbox_id"].as_str().unwrap().to_string(),
-                )
-            })
-            .collect();
-        records.sort_unstable();
-        records
-    }
-
-    async fn status_idle_sessions(status_path: &std::path::Path) -> Vec<String> {
-        status_idle_sessions_and_active_runs(status_path).await.0
-    }
-
-    async fn wait_status_idle_sessions_and_active_runs(
-        status_path: &std::path::Path,
-        expected_idle_sessions: &[&str],
-        expected_active_runs: &[String],
-        timeout: Duration,
-    ) {
-        let mut expected_idle_sessions: Vec<String> = expected_idle_sessions
-            .iter()
-            .map(|session| (*session).to_string())
-            .collect();
-        expected_idle_sessions.sort_unstable();
-        let mut expected_active_runs = expected_active_runs.to_vec();
-        expected_active_runs.sort_unstable();
-        let deadline = tokio::time::Instant::now() + timeout;
-        loop {
-            let Some((idle_sessions, active_runs)) =
-                status_idle_sessions_and_active_runs_if_exists(status_path).await
-            else {
-                assert!(
-                    tokio::time::Instant::now() < deadline,
-                    "status file {} was not written within {timeout:?}",
-                    status_path.display(),
-                );
-                tokio::time::sleep(Duration::from_millis(10)).await;
-                continue;
-            };
-            if idle_sessions == expected_idle_sessions && active_runs == expected_active_runs {
-                return;
-            }
-            assert!(
-                tokio::time::Instant::now() < deadline,
-                "status did not reach expected idle={expected_idle_sessions:?} active={expected_active_runs:?} within {timeout:?} (actual idle={idle_sessions:?} active={active_runs:?})",
-            );
-            tokio::time::sleep(Duration::from_millis(10)).await;
-        }
-    }
-
-    async fn status_idle_sessions_and_active_runs_if_exists(
-        status_path: &std::path::Path,
-    ) -> Option<(Vec<String>, Vec<String>)> {
-        match tokio::fs::try_exists(status_path).await {
-            Ok(true) => Some(status_idle_sessions_and_active_runs(status_path).await),
-            Ok(false) => None,
-            Err(err) => panic!(
-                "failed to check status file {}: {err}",
-                status_path.display()
-            ),
-        }
-    }
-
-    async fn publish_idle_status(pool: &SharedIdlePool, status: &StatusTracker) {
-        let snapshot = pool.lock().await.status_snapshot();
-        assert!(
-            status
-                .set_idle_info_at_revision(snapshot.revision, snapshot.idle_vms)
-                .await
-        );
-    }
-
-    async fn wait_status_idle_empty_with_active_run(
-        status_path: &std::path::Path,
-        run_id: RunId,
-        timeout: Duration,
-    ) {
-        let expected = run_id.to_string();
-        let deadline = tokio::time::Instant::now() + timeout;
-        loop {
-            let (idle_sessions, active_runs) =
-                status_idle_sessions_and_active_runs(status_path).await;
-            if idle_sessions.is_empty() && active_runs.iter().any(|id| id == &expected) {
-                return;
-            }
-            assert!(
-                tokio::time::Instant::now() < deadline,
-                "status did not atomically clear idle_vms and add active run {expected} within {timeout:?} (idle: {idle_sessions:?}, active: {active_runs:?})",
-            );
-            tokio::time::sleep(Duration::from_millis(10)).await;
-        }
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 10: Successful job parks VM in idle pool
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn successful_job_parks_in_idle_pool() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let budget = Arc::clone(&config.budget);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-park")),
-        );
-
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(completion.is_some(), "job should complete");
-        assert_eq!(completion.unwrap().exit_code, 0);
-
-        // Give park notification time to propagate.
-        tokio::time::sleep(Duration::from_millis(100)).await;
-
-        // VM should be parked in idle pool, holding budget.
-        {
-            let pool = idle_pool.lock().await;
-            assert_eq!(pool.len(), 1, "VM should be parked");
-            assert!(
-                pool.held_sessions().contains(&"sess-park".to_string()),
-                "parked session should be sess-park"
-            );
-        }
-        let (_, _, count) = budget.allocated();
-        assert_eq!(count, 1, "parked VM should hold budget");
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 11: Job without session destroys sandbox (no parking)
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn job_without_session_destroys_sandbox() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let budget = Arc::clone(&config.budget);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        // No resume_session → no session_id → no parking.
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(completion.is_some(), "job should complete");
-        assert_eq!(completion.unwrap().exit_code, 0);
-
-        // The active budget lease is dropped after provider.complete() in the
-        // spawned task, so wait_completion returning doesn't guarantee it has
-        // executed yet.
-        // Poll until budget is fully released rather than using a fixed sleep.
-        wait_budget_count(&budget, 0, Duration::from_secs(5)).await;
-
-        // No parking — pool empty, budget fully released.
-        assert_eq!(idle_pool.lock().await.len(), 0, "pool should be empty");
-
-        shutdown(&env, run_handle).await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn active_destroy_panic_still_reports_completion_and_releases_budget() {
-        let (config, env) = build_mock_run_config_with_runtime(
-            test_profiles(),
-            8,
-            32768,
-            4,
-            MockJobProvider::new,
-            Box::new(PanickingDestroyRuntime),
-            "http://localhost:0",
-        );
-        let budget = Arc::clone(&config.budget);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(
-            completion.is_some(),
-            "destroy panic must not skip provider.complete"
-        );
-        assert_eq!(completion.unwrap().exit_code, 0);
-
-        wait_budget_count(&budget, 0, Duration::from_secs(5)).await;
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 12: Park notification triggers immediate heartbeat
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn park_triggers_immediate_heartbeat() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let run_handle = tokio::spawn(run(config));
-
-        // Snapshot the heartbeat count before pushing a job. The first
-        // interval tick is now deferred by one period (`interval_at`), so
-        // `before` is typically 0; the 100 ms settle time just lets the
-        // main loop reach its idle select state.
-        tokio::time::sleep(Duration::from_millis(100)).await;
-        let before = env.handle.heartbeat_count();
-
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-hb")),
-        );
-
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(completion.is_some(), "job should complete");
-
-        // Give park notification time to trigger heartbeat.
-        tokio::time::sleep(Duration::from_millis(100)).await;
-        let after = env.handle.heartbeat_count();
-        assert!(
-            after > before,
-            "park should trigger at least one heartbeat (before={before}, after={after})"
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 13: Session affinity reuses idle VM
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn session_affinity_reuses_idle_vm() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let budget = Arc::clone(&config.budget);
-
-        // Pre-seed: park a VM for session "sess-reuse" with matching profile.
-        let seeded_sandbox_id =
-            seed_idle_pool(&idle_pool, &budget, "sess-reuse", "vm0/default", 2, 4096).await;
-        assert_eq!(budget.allocated().2, 1, "seeded entry holds budget");
-
-        let run_handle = tokio::spawn(run(config));
-
-        // Push job for same session — should reuse the idle VM.
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-reuse")),
-        );
-
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(completion.is_some(), "job should complete");
-        let completion = completion.unwrap();
-        assert_eq!(completion.exit_code, 0);
-        assert_eq!(
-            completion.reuse_result,
-            Some(SandboxReuseResult::Reused),
-            "reuse_result should be Reused"
-        );
-        assert_eq!(
-            completion.sandbox_id,
-            Some(seeded_sandbox_id),
-            "reused completion should carry the seeded sandbox id"
-        );
-
-        tokio::time::sleep(Duration::from_millis(100)).await;
-
-        // After reuse + re-park: pool should still have 1 entry, budget count=1.
-        {
-            let pool = idle_pool.lock().await;
-            assert_eq!(pool.len(), 1, "VM should be re-parked after reuse");
-        }
-        assert_eq!(
-            budget.allocated().2,
-            1,
-            "budget should remain at 1 (reused, not additive)"
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn reuse_take_clears_idle_status_while_job_is_active() {
-        let gate = Arc::new(tokio::sync::Notify::new());
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::with_wait_exit_gate(
-            Arc::clone(&gate),
-        ));
-        let (config, env) =
-            mock_run_config_with_overrides(test_profiles(), 8, 32768, 4, Arc::clone(&overrides));
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let budget = Arc::clone(&config.budget);
-        let status = Arc::clone(&config.status);
-        let status_path = env._temp_dir.path().join("status.json");
-
-        let _seeded_sandbox_id = seed_idle_pool_with_overrides(
-            &idle_pool,
-            &budget,
-            &overrides,
-            "sess-reuse-status",
-            "vm0/default",
-            2,
-            4096,
-        )
-        .await;
-        let snapshot = idle_pool.lock().await.status_snapshot();
-        assert!(
-            status
-                .set_idle_info_at_revision(snapshot.revision, snapshot.idle_vms)
-                .await
-        );
-        assert_eq!(
-            status_idle_sessions(&status_path).await,
-            vec!["sess-reuse-status".to_string()],
-            "pre-run status should list the seeded idle VM",
-        );
-
-        let run_handle = tokio::spawn(run(config));
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-reuse-status")),
-        );
-
-        wait_idle_pool_len(&idle_pool, 0, Duration::from_secs(5)).await;
-        wait_status_idle_empty_with_active_run(&status_path, run_id, Duration::from_secs(5)).await;
-
-        gate.notify_one();
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(completion.is_some(), "reused job should complete");
-        assert_eq!(
-            completion.unwrap().reuse_result,
-            Some(SandboxReuseResult::Reused),
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 13b: Job with no session reports NoSessionId
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn job_without_session_reports_no_session_id() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-
-        let run_handle = tokio::spawn(run(config));
-
-        // No resume_session → NoSessionId branch.
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(completion.is_some(), "job should complete");
-        let completion = completion.unwrap();
-        assert_eq!(completion.exit_code, 0);
-        assert_eq!(
-            completion.reuse_result,
-            Some(SandboxReuseResult::NoSessionId),
-        );
-        assert!(
-            completion.sandbox_id.is_some(),
-            "fresh create still allocates a sandbox id",
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 14: Profile mismatch destroys stale and creates new
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn profile_mismatch_destroys_stale_vm() {
-        let (config, env) = mock_run_config(two_profiles(), 16, 32768, 4);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let budget = Arc::clone(&config.budget);
-
-        // Pre-seed: park a "vm0/default" (2vcpu) VM for session "sess-mm".
-        seed_idle_pool(&idle_pool, &budget, "sess-mm", "vm0/default", 2, 4096).await;
-
-        let run_handle = tokio::spawn(run(config));
-
-        // Push job for "vm0/large" (4vcpu) with same session — profile mismatch.
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/large",
-            Some(context_with_session(run_id, "sess-mm")),
-        );
-
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(completion.is_some(), "job should complete");
-        let completion = completion.unwrap();
-        assert_eq!(completion.exit_code, 0);
-        assert_eq!(
-            completion.reuse_result,
-            Some(SandboxReuseResult::ProfileMismatch),
-            "reuse_result should be ProfileMismatch when profile differs"
-        );
-        assert!(
-            completion.sandbox_id.is_some(),
-            "freshly created sandbox still reports its id"
-        );
-
-        // Stale VM destruction runs in a background destroy_task. Poll until
-        // its budget is released rather than using a fixed sleep.
-        // Expected: stale 2vcpu released, new 4vcpu held → count=1.
-        wait_budget_count(&budget, 1, Duration::from_secs(5)).await;
-
-        {
-            let pool = idle_pool.lock().await;
-            assert_eq!(pool.len(), 1, "new VM should be parked");
-        }
-        let (alloc_vcpu, alloc_mem, alloc_count) = budget.allocated();
-        assert_eq!(alloc_count, 1, "only new VM should hold budget");
-        assert_eq!(alloc_vcpu, 4, "new VM is vm0/large (4 vcpu)");
-        assert_eq!(alloc_mem, 8192, "new VM is vm0/large (8192 MB)");
-
-        shutdown(&env, run_handle).await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn profile_mismatch_status_switches_from_idle_to_active_while_job_runs() {
-        let gate = Arc::new(tokio::sync::Notify::new());
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::with_wait_exit_gate(
-            Arc::clone(&gate),
-        ));
-        let (config, env) = mock_run_config_with_overrides(two_profiles(), 16, 32768, 4, overrides);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let budget = Arc::clone(&config.budget);
-        let status = Arc::clone(&config.status);
-        let status_path = env._temp_dir.path().join("status.json");
-
-        seed_idle_pool(
-            &idle_pool,
-            &budget,
-            "sess-mm-status",
-            "vm0/default",
-            2,
-            4096,
-        )
-        .await;
-        publish_idle_status(&idle_pool, &status).await;
-        assert_eq!(
-            status_idle_sessions(&status_path).await,
-            vec!["sess-mm-status".to_string()],
-            "pre-run status should list the stale idle VM",
-        );
-
-        let run_handle = tokio::spawn(run(config));
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/large",
-            Some(context_with_session(run_id, "sess-mm-status")),
-        );
-
-        wait_idle_pool_len(&idle_pool, 0, Duration::from_secs(5)).await;
-        wait_status_idle_empty_with_active_run(&status_path, run_id, Duration::from_secs(5)).await;
-
-        gate.notify_one();
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await
-            .expect("fresh-create job should still complete");
-        assert_eq!(
-            completion.reuse_result,
-            Some(SandboxReuseResult::ProfileMismatch),
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 15: Cleanup tick evicts expired idle entries
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn cleanup_tick_evicts_expired_entries() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let budget = Arc::clone(&config.budget);
-
-        // Pre-seed: park an entry that is already expired (400s old, 300s timeout).
-        seed_idle_pool_expired(&idle_pool, &budget, "sess-exp", "vm0/default", 2, 4096).await;
-        assert_eq!(
-            idle_pool.lock().await.len(),
-            1,
-            "should have 1 seeded entry"
-        );
-        assert_eq!(budget.allocated().2, 1, "seeded entry holds budget");
-
-        let run_handle = tokio::spawn(run(config));
-
-        // Advance past the first cleanup tick (every 10s).
-        // The tick interval fires once immediately (at t=0), but the entry
-        // was just inserted so it may not be expired yet from Instant::now()'s
-        // perspective. Advance 11s to ensure at least one full tick fires.
-        tokio::time::sleep(Duration::from_secs(11)).await;
-
-        // Eviction spawns a destroy_task that releases the idle entry lease.
-        // Poll until it completes.
-        wait_budget_count(&budget, 0, Duration::from_secs(5)).await;
-
-        let pool_len = idle_pool.lock().await.len();
-        assert_eq!(pool_len, 0, "expired entry should be evicted");
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 16: Budget exhausted → evict idle VM → admit new job
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn budget_exhausted_evicts_idle_and_admits_job() {
-        // Budget: exactly 1 default job (2 vcpu, 4096 MB).
-        let (config, env) = mock_run_config(test_profiles(), 2, 4096, 2);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let budget = Arc::clone(&config.budget);
-
-        // Pre-seed: idle VM fills the entire budget.
-        seed_idle_pool(&idle_pool, &budget, "sess-evict", "vm0/default", 2, 4096).await;
-        assert!(
-            !budget.can_afford(2, 4096),
-            "budget should be exhausted after seeding"
-        );
-
-        let run_handle = tokio::spawn(run(config));
-
-        // Push new job — budget is full, but idle pool has an entry to evict.
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(
-            completion.is_some(),
-            "job should complete after idle VM eviction frees budget"
-        );
-        assert_eq!(completion.unwrap().exit_code, 0);
-
-        shutdown(&env, run_handle).await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn budget_exhausted_reclaims_expired_before_oldest_idle() {
-        let (config, env) = mock_run_config(two_profiles(), 6, 12288, 3);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let budget = Arc::clone(&config.budget);
-        let status_path = env._temp_dir.path().join("status.json");
-        let now = std::time::Instant::now();
-
-        seed_idle_pool_with_timing(
-            &idle_pool,
-            &budget,
-            TestParkCandidateSpec {
-                session_id: "sess-old-active",
-                profile_name: "vm0/default",
-                vcpu: 2,
-                memory_mb: 4096,
-                parked_at: now - Duration::from_secs(100),
-                idle_timeout: Duration::from_secs(300),
-            },
-        )
-        .await;
-        seed_idle_pool_with_timing(
-            &idle_pool,
-            &budget,
-            TestParkCandidateSpec {
-                session_id: "sess-expired-newer",
-                profile_name: "vm0/large",
-                vcpu: 4,
-                memory_mb: 8192,
-                parked_at: now - Duration::from_secs(10),
-                idle_timeout: Duration::from_secs(1),
-            },
-        )
-        .await;
-        assert!(
-            !budget.can_afford(2, 4096),
-            "seeded idle entries should exhaust budget"
-        );
-
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(
-            completion.is_some(),
-            "job should complete after expired idle reclaim frees budget"
-        );
-        assert_eq!(completion.unwrap().exit_code, 0);
-
-        wait_budget_count(&budget, 1, Duration::from_secs(5)).await;
-
-        let sessions = idle_pool.lock().await.held_sessions();
-        assert_eq!(
-            sessions,
-            vec!["sess-old-active".to_string()],
-            "expired idle entry should be reclaimed before oldest active entry"
-        );
-        assert_eq!(
-            status_idle_sessions(&status_path).await,
-            vec!["sess-old-active".to_string()],
-            "status.json should reflect the remaining idle VM"
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn budget_exhausted_evicts_oldest_when_expired_reclaim_insufficient() {
-        let (config, env) = mock_run_config(two_profiles(), 7, 13312, 3);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let budget = Arc::clone(&config.budget);
-        let status_path = env._temp_dir.path().join("status.json");
-        let now = std::time::Instant::now();
-
-        seed_idle_pool_with_timing(
-            &idle_pool,
-            &budget,
-            TestParkCandidateSpec {
-                session_id: "sess-old-active",
-                profile_name: "vm0/large",
-                vcpu: 4,
-                memory_mb: 8192,
-                parked_at: now - Duration::from_secs(100),
-                idle_timeout: Duration::from_secs(300),
-            },
-        )
-        .await;
-        seed_idle_pool_with_timing(
-            &idle_pool,
-            &budget,
-            TestParkCandidateSpec {
-                session_id: "sess-new-active",
-                profile_name: "vm0/default",
-                vcpu: 2,
-                memory_mb: 4096,
-                parked_at: now - Duration::from_secs(50),
-                idle_timeout: Duration::from_secs(300),
-            },
-        )
-        .await;
-        seed_idle_pool_with_timing(
-            &idle_pool,
-            &budget,
-            TestParkCandidateSpec {
-                session_id: "sess-expired-small",
-                profile_name: "vm0/default",
-                // Intentionally smaller than the current min profile. With
-                // only profile-sized entries, releasing one expired VM is
-                // already enough to admit the min profile; this pins the
-                // fallback loop for stale/non-current idle footprints.
-                vcpu: 1,
-                memory_mb: 1024,
-                parked_at: now - Duration::from_secs(10),
-                idle_timeout: Duration::from_secs(1),
-            },
-        )
-        .await;
-        assert!(
-            !budget.can_afford(2, 4096),
-            "seeded idle entries should exhaust budget"
-        );
-
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(
-            completion.is_some(),
-            "job should complete after expired reclaim plus oldest eviction"
-        );
-        assert_eq!(completion.unwrap().exit_code, 0);
-
-        wait_budget_count(&budget, 1, Duration::from_secs(5)).await;
-
-        let sessions = idle_pool.lock().await.held_sessions();
-        assert_eq!(
-            sessions,
-            vec!["sess-new-active".to_string()],
-            "expired entry and oldest active entry should be reclaimed"
-        );
-        assert_eq!(
-            status_idle_sessions(&status_path).await,
-            vec!["sess-new-active".to_string()],
-            "status.json should reflect only the remaining idle VM"
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn budget_pressure_eviction_clears_status_json_idle_vms() {
-        let (config, env) = mock_run_config(test_profiles(), 2, 4096, 2);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let budget = Arc::clone(&config.budget);
-        let status_path = env._temp_dir.path().join("status.json");
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-pressure-status")),
-        );
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(completion.is_some(), "job should complete and park");
-        assert_eq!(completion.unwrap().exit_code, 0);
-
-        // The single parked VM fills the whole budget, so the Running loop's
-        // pressure path evicts it even without another pending job.
-        wait_budget_count(&budget, 0, Duration::from_secs(5)).await;
-        assert_eq!(idle_pool.lock().await.len(), 0, "idle pool should be empty");
-        assert!(
-            status_idle_sessions(&status_path).await.is_empty(),
-            "status.json should clear the pressure-evicted idle VM"
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 17: Shutdown drains idle pool and releases budget
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn shutdown_drains_idle_pool() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let budget = Arc::clone(&config.budget);
-
-        // Pre-seed: two idle entries holding budget.
-        seed_idle_pool(&idle_pool, &budget, "sess-drain-1", "vm0/default", 2, 4096).await;
-        seed_idle_pool(&idle_pool, &budget, "sess-drain-2", "vm0/default", 2, 4096).await;
-        assert_eq!(idle_pool.lock().await.len(), 2);
-        assert_eq!(budget.allocated().2, 2);
-
-        let run_handle = tokio::spawn(run(config));
-
-        // Immediately shutdown — drain should destroy all idle entries.
-        shutdown(&env, run_handle).await;
-
-        // After shutdown: pool empty, budget fully released.
-        assert_eq!(idle_pool.lock().await.len(), 0, "pool should be drained");
-        let (_, _, count) = budget.allocated();
-        assert_eq!(count, 0, "all budget should be released after drain");
-    }
-
-    /// Active soft drain closes parking for successful jobs that complete
-    /// before SIGUSR2 resume. The sandbox is destroyed and budget is released
-    /// instead of late-parking into an already-drained pool.
-    #[tokio::test]
-    async fn job_completing_during_active_draining_is_not_parked() {
-        let gate = Arc::new(tokio::sync::Notify::new());
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::with_wait_exit_gate(
-            Arc::clone(&gate),
-        ));
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 32768, 4, overrides);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let budget = Arc::clone(&config.budget);
-        let run_handle = tokio::spawn(run(config));
-
-        // Claim a gated job with a reusable session while Running.
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-late-park")),
-        );
-        let _token = wait_cancel_token(&env.cancel_tokens, run_id, Duration::from_secs(5)).await;
-
-        // Enter Draining. The Draining path drains an empty pool and waits for the
-        // gated job.
-        env.drain();
-        wait_parking_state(
-            &idle_pool,
-            ParkingState::SoftDraining,
-            Duration::from_secs(5),
-        )
-        .await;
-        assert_eq!(
-            idle_pool.lock().await.len(),
-            0,
-            "Draining mode should have drained an empty pool",
-        );
-
-        // Release the gate while still Draining: parking is closed, so the
-        // successful job destroys its sandbox instead of parking it.
-        gate.notify_one();
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(c.is_some(), "job should complete");
-        assert_eq!(c.unwrap().exit_code, 0);
-
-        assert_eq!(
-            idle_pool.lock().await.len(),
-            0,
-            "active draining must reject post-job parking",
-        );
-
-        // Draining mode observes jobs.is_empty → auto-Stop → teardown.
-        match tokio::time::timeout(Duration::from_secs(5), run_handle).await {
-            Ok(Ok(Ok(()))) => {}
-            Ok(Ok(Err(e))) => panic!("run() returned error: {e}"),
-            Ok(Err(e)) => panic!("task panicked: {e}"),
-            Err(_) => panic!("natural drain should exit within 5s"),
-        }
-
-        // Leak proof: pool empty, budget fully released.
-        assert_eq!(
-            idle_pool.lock().await.len(),
-            0,
-            "teardown must leave no idle VM",
-        );
-        assert_eq!(
-            budget.allocated().2,
-            0,
-            "budget must be fully released (no held entries, no stray reservations)",
-        );
-    }
-
-    /// Regression for #11162: once SIGUSR2 has logically resumed the runner,
-    /// parking is open even if the main loop has not yet processed the Running
-    /// tick. The silent mode flip keeps the main loop in the pre-ack window
-    /// deterministically.
-    #[tokio::test]
-    async fn soft_drain_resume_opens_parking_before_running_ack() {
-        let gate = Arc::new(tokio::sync::Notify::new());
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::with_wait_exit_gate(
-            Arc::clone(&gate),
-        ));
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 32768, 4, overrides);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let budget = Arc::clone(&config.budget);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-soft-resume-race")),
-        );
-        let _token = wait_cancel_token(&env.cancel_tokens, run_id, Duration::from_secs(5)).await;
-
-        env.drain();
-        wait_parking_state(
-            &idle_pool,
-            ParkingState::SoftDraining,
-            Duration::from_secs(5),
-        )
-        .await;
-
-        // Simulate SIGUSR2's ordering while suppressing the watch wake: open
-        // parking first, then make Running visible without letting the main
-        // loop run its top-of-loop Running branch.
-        env.parking_gate.open_after_soft_drain();
-        env.mode_tx.send_if_modified(|mode| {
-            *mode = RunnerMode::Running;
-            false
-        });
-        assert_eq!(*env.mode_tx.borrow(), RunnerMode::Running);
-
-        gate.notify_one();
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(c.is_some(), "job should complete after logical resume");
-        assert_eq!(c.unwrap().exit_code, 0);
-
-        assert_eq!(
-            idle_pool.lock().await.len(),
-            1,
-            "job should park even before the main loop acknowledges Running",
-        );
-        assert_eq!(
-            budget.allocated().2,
-            1,
-            "parked VM should retain its budget lease",
-        );
-
-        env.trigger_stopping().await;
-        match tokio::time::timeout(Duration::from_secs(5), run_handle).await {
-            Ok(Ok(Ok(()))) => {}
-            Ok(Ok(Err(e))) => panic!("run() returned error: {e}"),
-            Ok(Err(e)) => panic!("task panicked: {e}"),
-            Err(_) => panic!("hard shutdown should exit within 5s"),
-        }
-    }
-
-    /// Regression (G2): on SIGTERM from Running, teardown's
-    /// `drain_idle_pool` is the *only* site that clears `idle_vms` in
-    /// `status.json` — Draining mode is skipped entirely. Pre-fix, the
-    /// stale list leaked into the final `"stopped"` snapshot.
-    #[tokio::test(start_paused = true)]
-    async fn shutdown_clears_idle_vms_in_status_json() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let status_path = env._temp_dir.path().join("status.json");
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let run_handle = tokio::spawn(run(config));
-
-        // Park a VM via a normal job → status.json records the idle VM.
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-status-clean")),
-        );
-        let _ = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        tokio::time::sleep(Duration::from_millis(100)).await;
-        assert_eq!(idle_pool.lock().await.len(), 1, "VM parked");
-
-        // Pre-shutdown sanity: status.json lists the idle VM.
-        let pre: serde_json::Value =
-            serde_json::from_str(&tokio::fs::read_to_string(&status_path).await.unwrap()).unwrap();
-        let pre_len = pre
-            .get("idle_vms")
-            .and_then(|v| v.as_array())
-            .map(|a| a.len())
-            .unwrap_or(0);
-        assert_eq!(pre_len, 1, "pre-shutdown status.json should list the VM");
-
-        // SIGTERM path: Draining mode is bypassed, so teardown's
-        // drain_idle_pool is the only site that can clear idle_vms.
-        env.trigger_stopping().await;
-        match tokio::time::timeout(Duration::from_secs(5), run_handle).await {
-            Ok(Ok(Ok(()))) => {}
-            Ok(Ok(Err(e))) => panic!("run() returned error: {e}"),
-            Ok(Err(e)) => panic!("task panicked: {e}"),
-            Err(_) => panic!("hard shutdown should exit within 5s"),
-        }
-
-        // Post-shutdown: mode=stopped, idle_vms empty/absent.
-        let post: serde_json::Value =
-            serde_json::from_str(&tokio::fs::read_to_string(&status_path).await.unwrap()).unwrap();
-        assert_eq!(post["mode"], "stopped");
-        let post_len = post
-            .get("idle_vms")
-            .and_then(|v| v.as_array())
-            .map(|a| a.len())
-            .unwrap_or(0);
-        assert_eq!(
-            post_len, 0,
-            "status.json idle_vms must be cleared after shutdown: {post}",
-        );
-    }
-
-    // -----------------------------------------------------------------------
-    // Test 19: Two sequential jobs for same session → take + reuse + re-park
-    //
-    // Exercises the full session affinity cycle: park → take → reuse → park.
-    // After two jobs the pool should have exactly 1 entry (the second job's
-    // VM) and the budget count should be 1.
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn sequential_same_session_reuse_cycle() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let budget = Arc::clone(&config.budget);
-        let run_handle = tokio::spawn(run(config));
-
-        // Job 1: parks VM for session "sess-seq".
-        let id1 = RunId::new_v4();
-        push_job(
-            &env,
-            id1,
-            "vm0/default",
-            Some(context_with_session(id1, "sess-seq")),
-        );
-        let c1 = env
-            .handle
-            .wait_completion(id1, Duration::from_secs(5))
-            .await;
-        assert!(c1.is_some(), "job 1 should complete");
-        tokio::time::sleep(Duration::from_millis(100)).await;
-        assert_eq!(idle_pool.lock().await.len(), 1, "job 1 VM should be parked");
-
-        // Job 2: same session → take → reuse → re-park.
-        let id2 = RunId::new_v4();
-        push_job(
-            &env,
-            id2,
-            "vm0/default",
-            Some(context_with_session(id2, "sess-seq")),
-        );
-        let c2 = env
-            .handle
-            .wait_completion(id2, Duration::from_secs(5))
-            .await;
-        assert!(c2.is_some(), "job 2 should complete");
-        tokio::time::sleep(Duration::from_millis(100)).await;
-
-        assert_eq!(
-            idle_pool.lock().await.len(),
-            1,
-            "pool should have 1 entry after two sequential jobs"
-        );
-        assert_eq!(budget.allocated().2, 1, "only one VM should hold budget");
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // =======================================================================
-    // Tests 20-22: Edge cases requiring MockSandboxOverrides
-    // =======================================================================
-
-    fn mock_run_config_with_overrides(
-        profiles: BTreeMap<String, config::ProfileConfig>,
-        budget_vcpu: u32,
-        budget_memory_mb: u32,
-        max_concurrent: usize,
-        overrides: Arc<sandbox_mock::MockSandboxOverrides>,
-    ) -> (RunConfig, MockRunEnv) {
-        build_mock_run_config_with_runtime(
-            profiles,
-            budget_vcpu,
-            budget_memory_mb,
-            max_concurrent,
-            MockJobProvider::new,
-            Box::new(MockSandboxRuntime::with_overrides(overrides)),
-            "http://localhost:0",
-        )
-    }
-
-    async fn wait_cancel_token(
-        tokens: &Arc<tokio::sync::Mutex<HashMap<RunId, CancellationToken>>>,
-        run_id: RunId,
-        timeout: Duration,
-    ) -> CancellationToken {
-        let deadline = tokio::time::Instant::now() + timeout;
-        loop {
-            if let Some(token) = tokens.lock().await.get(&run_id).cloned() {
-                return token;
-            }
-            assert!(
-                tokio::time::Instant::now() < deadline,
-                "cancel token for {run_id} not found within {timeout:?}",
-            );
-            tokio::time::sleep(Duration::from_millis(10)).await;
-        }
-    }
-
-    async fn wait_cancel_token_removed(
-        tokens: &Arc<tokio::sync::Mutex<HashMap<RunId, CancellationToken>>>,
-        run_id: RunId,
-        timeout: Duration,
-    ) {
-        let deadline = tokio::time::Instant::now() + timeout;
-        loop {
-            if !tokens.lock().await.contains_key(&run_id) {
-                return;
-            }
-            assert!(
-                tokio::time::Instant::now() < deadline,
-                "cancel token for {run_id} still present after {timeout:?}",
-            );
-            tokio::time::sleep(Duration::from_millis(10)).await;
-        }
-    }
-
-    /// Test 20: Failed job with session context is not parked.
-    ///
-    /// When `wait_exit` returns a non-zero exit code, `spawn_job()` skips
-    /// parking (because `exit_code == 0` is false) and destroys the sandbox.
-    #[tokio::test(start_paused = true)]
-    async fn failed_job_with_session_not_parked() {
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::with_wait_exit_code(1));
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 4, 8192, 4, overrides);
-        let budget = Arc::clone(&config.budget);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-fail")),
-        );
-
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(c.is_some(), "job should complete");
-        assert_eq!(c.unwrap().exit_code, 1);
-
-        wait_budget_count(&budget, 0, Duration::from_secs(2)).await;
-        assert_eq!(idle_pool.lock().await.len(), 0, "failed job must not park");
-
-        shutdown(&env, run_handle).await;
-    }
-
-    /// Test 21: Cancelled job is not parked.
-    ///
-    /// `wait_exit_gate` blocks the agent execution. The test cancels the job
-    /// via the cancel token, causing `select!` in the executor to take the
-    /// cancellation branch. `job_cancel.is_cancelled()` is true, so
-    /// `parkable_session` is `None` → sandbox destroyed.
-    #[tokio::test(start_paused = true)]
-    async fn cancelled_job_not_parked() {
-        let gate = Arc::new(tokio::sync::Notify::new());
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::with_wait_exit_gate(
-            gate,
-        ));
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 4, 8192, 4, overrides);
-        let budget = Arc::clone(&config.budget);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let cancel_tokens = Arc::clone(&config.cancel_tokens);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-cancel")),
-        );
-
-        // Wait for the job to be claimed (cancel token inserted).
-        let token = wait_cancel_token(&cancel_tokens, run_id, Duration::from_secs(5)).await;
-        // Cancel the job — executor's select! takes the cancelled branch.
-        token.cancel();
-
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(c.is_some(), "cancelled job should still complete");
-        let c = c.unwrap();
-        assert_eq!(c.exit_code, 137, "cancellation yields synthetic SIGKILL");
-        assert_eq!(c.error.as_deref(), Some("cancelled by user"));
-
-        wait_budget_count(&budget, 0, Duration::from_secs(2)).await;
-        assert_eq!(
-            idle_pool.lock().await.len(),
-            0,
-            "cancelled job must not park"
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    /// Test 22: `ParkResult::Replaced` via `guest_session_id`.
-    ///
-    /// A first-run job (no `resume_session`) reads a CLI-generated session ID
-    /// from the guest filesystem. When that session already has an entry in
-    /// the idle pool, `pool.park()` returns `Replaced(old)`, the old VM is
-    /// destroyed, and the new VM takes its place.
-    #[tokio::test(start_paused = true)]
-    async fn park_evicts_via_guest_session_id() {
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::new());
-        overrides.add_exec_matcher(sandbox_mock::ExecMatcher {
-            pattern: "cat /tmp/vm0-session-".into(),
-            exit_code: 0,
-            stdout: b"sess-evict".to_vec(),
-            stderr: Vec::new(),
-        });
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 16384, 4, overrides);
-        let budget = Arc::clone(&config.budget);
-        let idle_pool = Arc::clone(&config.idle_pool);
-
-        // Pre-seed idle pool with session "sess-evict".
-        seed_idle_pool(&idle_pool, &budget, "sess-evict", "vm0/default", 2, 4096).await;
-        assert_eq!(budget.allocated().2, 1, "pre-seeded entry holds budget");
-
-        let run_handle = tokio::spawn(run(config));
-
-        // Push job WITHOUT resume_session — first run, no session context.
-        // read_guest_session_id() will be called and return "sess-evict"
-        // via the exec matcher.
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(c.is_some(), "job should complete");
-        assert_eq!(c.unwrap().exit_code, 0);
-
-        // After eviction: old entry destroyed + old budget released,
-        // new entry parked + new budget held → net count = 1.
-        wait_budget_count(&budget, 1, Duration::from_secs(2)).await;
-        let pool = idle_pool.lock().await;
-        assert_eq!(pool.len(), 1, "pool should have the newly parked entry");
-        assert_eq!(
-            pool.held_sessions(),
-            vec!["sess-evict"],
-            "parked session should match guest_session_id"
-        );
-        drop(pool);
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Tests 23-25: park / unpark idle-transition orchestration (#9102)
-    // -----------------------------------------------------------------------
-
-    /// Two sequential jobs on the same session produce park=2 / unpark=1:
-    /// the first job's post-exit park, plus the second job's take (unpark)
-    /// and post-exit re-park. Verifies the full reuse cycle drives the
-    /// new trait hooks symmetrically.
-    #[tokio::test(start_paused = true)]
-    async fn reuse_cycle_invokes_park_and_unpark_symmetrically() {
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::new());
-        let counter = Arc::clone(&overrides);
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 32768, 4, overrides);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let run_handle = tokio::spawn(run(config));
-
-        // Job 1: fresh create → run → park.
-        let id1 = RunId::new_v4();
-        push_job(
-            &env,
-            id1,
-            "vm0/default",
-            Some(context_with_session(id1, "sess-reuse-cycle")),
-        );
-        assert!(
-            env.handle
-                .wait_completion(id1, Duration::from_secs(5))
-                .await
-                .is_some()
-        );
-        tokio::time::sleep(Duration::from_millis(100)).await;
-        assert_eq!(counter.park_call_count(), 1);
-        assert_eq!(counter.unpark_call_count(), 0);
-
-        // Job 2: same session → take (unpark) → run → re-park.
-        let id2 = RunId::new_v4();
-        push_job(
-            &env,
-            id2,
-            "vm0/default",
-            Some(context_with_session(id2, "sess-reuse-cycle")),
-        );
-        assert!(
-            env.handle
-                .wait_completion(id2, Duration::from_secs(5))
-                .await
-                .is_some()
-        );
-        tokio::time::sleep(Duration::from_millis(100)).await;
-        assert_eq!(
-            counter.park_call_count(),
-            2,
-            "park() should fire once per job"
-        );
-        assert_eq!(
-            counter.unpark_call_count(),
-            1,
-            "unpark() should fire only for the reused job"
-        );
-        assert_eq!(idle_pool.lock().await.len(), 1);
-
-        shutdown(&env, run_handle).await;
-    }
-
-    /// A successful job with a session triggers `Sandbox::park()` exactly once
-    /// when the VM is handed off to the idle pool.
-    #[tokio::test(start_paused = true)]
-    async fn park_called_when_vm_enters_idle_pool() {
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::new());
-        let counter = Arc::clone(&overrides);
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 16384, 4, overrides);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-park-hook")),
-        );
-
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(c.is_some(), "job should complete");
-
-        // Park notification + park() are called from the post-job task; give
-        // them a moment to run before asserting.
-        tokio::time::sleep(Duration::from_millis(100)).await;
-
-        assert_eq!(
-            counter.park_call_count(),
-            1,
-            "park() should have been called exactly once"
-        );
-        assert_eq!(
-            counter.unpark_call_count(),
-            0,
-            "unpark() must not be called for a fresh park"
-        );
-        assert_eq!(idle_pool.lock().await.len(), 1, "VM should be parked");
-
-        shutdown(&env, run_handle).await;
-    }
-
-    /// When `Sandbox::park()` returns an error, the runner falls back to
-    /// `stop_and_destroy_sandbox` and does NOT insert into the idle pool.
-    #[tokio::test(start_paused = true)]
-    async fn park_failure_destroys_sandbox_and_skips_pool() {
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::new());
-        overrides.push_park_result(Err(sandbox::SandboxError::IdleTransition {
-            transition: sandbox::SandboxIdleTransition::Park,
-            message: "simulated balloon failure".into(),
-        }));
-        let counter = Arc::clone(&overrides);
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 16384, 4, overrides);
-        let budget = Arc::clone(&config.budget);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-park-fail")),
-        );
-
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(c.is_some(), "job should complete normally");
-        assert_eq!(c.unwrap().exit_code, 0);
-
-        // park failure → destroy → budget fully released, pool empty.
-        wait_budget_count(&budget, 0, Duration::from_secs(2)).await;
-        assert_eq!(
-            idle_pool.lock().await.len(),
-            0,
-            "park failure must NOT insert into pool"
-        );
-        assert_eq!(
-            counter.park_call_count(),
-            1,
-            "park() should have been attempted exactly once"
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn park_panic_destroys_sandbox_reports_completion_and_releases_budget() {
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::new());
-        overrides.push_park_panic("simulated park panic");
-        let counter = Arc::clone(&overrides);
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 16384, 4, overrides);
-        let budget = Arc::clone(&config.budget);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-park-panic")),
-        );
-
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(c.is_some(), "park panic must not skip provider.complete");
-        assert_eq!(c.unwrap().exit_code, 0);
-
-        wait_budget_count(&budget, 0, Duration::from_secs(2)).await;
-        assert_eq!(idle_pool.lock().await.len(), 0);
-        assert_eq!(counter.park_call_count(), 1);
-
-        shutdown(&env, run_handle).await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn create_failure_completes_and_cleans_run_state() {
-        let (config, env) = build_mock_run_config_with_runtime(
-            test_profiles(),
-            8,
-            16384,
-            4,
-            MockJobProvider::new,
-            Box::new(FailingCreateRuntime),
-            "http://localhost:0",
-        );
-        let budget = Arc::clone(&config.budget);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let cancel_tokens = Arc::clone(&config.cancel_tokens);
-        let status_path = env._temp_dir.path().join("status.json");
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await
-            .expect("create failure should still report completion");
-        assert_eq!(c.exit_code, 1);
-        let error = c.error.expect("create failure should report an error");
-        assert!(error.contains("create failed"), "got: {error}");
-
-        wait_budget_count(&budget, 0, Duration::from_secs(2)).await;
-        wait_cancel_token_removed(&cancel_tokens, run_id, Duration::from_secs(2)).await;
-        assert_eq!(idle_pool.lock().await.len(), 0);
-        let (_idle_sessions, active_runs) =
-            status_idle_sessions_and_active_runs(&status_path).await;
-        assert!(
-            active_runs.is_empty(),
-            "create failure should remove active run from status"
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn outer_job_panic_after_idle_pool_owned_cleans_token_and_active_status() {
-        let (mut config, env) = mock_run_config_with_overrides(
-            test_profiles(),
-            8,
-            16384,
-            4,
-            Arc::new(sandbox_mock::MockSandboxOverrides::new()),
-        );
-        config.outer_job_panic = Some(OuterJobPanicPoint::IdlePoolOwned);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let cancel_tokens = Arc::clone(&config.cancel_tokens);
-        let status_path = env._temp_dir.path().join("status.json");
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-outer-panic-idle")),
-        );
-
-        wait_idle_pool_len(&idle_pool, 1, Duration::from_secs(5)).await;
-        wait_cancel_token_removed(&cancel_tokens, run_id, Duration::from_secs(5)).await;
-        wait_status_idle_sessions_and_active_runs(
-            &status_path,
-            &["sess-outer-panic-idle"],
-            &[],
-            Duration::from_secs(5),
-        )
-        .await;
-        assert!(
-            !env.handle
-                .completions
-                .lock()
-                .unwrap()
-                .iter()
-                .any(|completion| completion.run_id == run_id),
-            "outer job panic must not synthesize provider completion"
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn outer_job_panic_active_unknown_reconciles_on_shutdown_final_scan() {
-        let (mut config, env) = mock_run_config_with_overrides(
-            test_profiles(),
-            8,
-            16384,
-            4,
-            Arc::new(sandbox_mock::MockSandboxOverrides::new()),
-        );
-        config.outer_job_panic = Some(OuterJobPanicPoint::ActiveOrUnknown);
-        config.orphan_reap_process_discovery = Some(OrphanReapProcessDiscovery {
-            firecrackers: Arc::new(Vec::new()),
-            incomplete_for_current_runner: false,
-        });
-        let cancel_tokens = Arc::clone(&config.cancel_tokens);
-        let status_path = env._temp_dir.path().join("status.json");
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        wait_cancel_token_removed(&cancel_tokens, run_id, Duration::from_secs(5)).await;
-        wait_status_idle_sessions_and_active_runs(
-            &status_path,
-            &[],
-            &[run_id.to_string()],
-            Duration::from_secs(5),
-        )
-        .await;
-
-        shutdown(&env, run_handle).await;
-        wait_status_idle_sessions_and_active_runs(&status_path, &[], &[], Duration::from_secs(5))
-            .await;
-        assert!(
-            !env.handle
-                .completions
-                .lock()
-                .unwrap()
-                .iter()
-                .any(|completion| completion.run_id == run_id),
-            "outer job panic must not synthesize provider completion"
-        );
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn outer_job_panic_after_destroy_completed_cleans_token_and_active_status() {
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::new());
-        let destroy_entered = Arc::new(tokio::sync::Notify::new());
-        let destroy_release = Arc::new(tokio::sync::Notify::new());
-        overrides.set_destroy_gate(Arc::clone(&destroy_entered), Arc::clone(&destroy_release));
-        let destroy_entered_wait = destroy_entered.notified();
-        tokio::pin!(destroy_entered_wait);
-        destroy_entered_wait.as_mut().enable();
-
-        let (mut config, env) =
-            mock_run_config_with_overrides(test_profiles(), 8, 16384, 4, overrides);
-        config.outer_job_panic = Some(OuterJobPanicPoint::DestroyCompleted);
-        let cancel_tokens = Arc::clone(&config.cancel_tokens);
-        let status_path = env._temp_dir.path().join("status.json");
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
-
-        tokio::time::timeout(Duration::from_secs(5), destroy_entered_wait)
-            .await
-            .expect("active destroy should start");
-        let _token = wait_cancel_token(&cancel_tokens, run_id, Duration::from_secs(5)).await;
-        destroy_release.notify_one();
-
-        wait_cancel_token_removed(&cancel_tokens, run_id, Duration::from_secs(5)).await;
-        wait_status_idle_sessions_and_active_runs(&status_path, &[], &[], Duration::from_secs(5))
-            .await;
-        assert!(
-            !env.handle
-                .completions
-                .lock()
-                .unwrap()
-                .iter()
-                .any(|completion| completion.run_id == run_id),
-            "outer job panic must not synthesize provider completion"
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn pool_full_rejected_vm_keeps_budget_until_destroy_and_completion() {
-        let hooks = BlockingDestroyHooks::new();
-        let (config, env) = build_mock_run_config_with_runtime(
-            test_profiles(),
-            8,
-            16384,
-            4,
-            MockJobProvider::new,
-            Box::new(BlockingDestroyRuntime {
-                hooks: hooks.clone(),
-            }),
-            "http://localhost:0",
-        );
-        let budget = Arc::clone(&config.budget);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        {
-            let mut pool = idle_pool.lock().await;
-            *pool = IdlePool::new(IdlePoolConfig {
-                default_timeout: Duration::from_secs(300),
-                max_idle: 1,
-            });
-        }
-        seed_idle_pool(&idle_pool, &budget, "sess-existing", "vm0/default", 2, 4096).await;
-        assert_eq!(budget.allocated().2, 1, "seeded idle entry holds budget");
-
-        let destroy_entered = hooks.entered.notified();
-        tokio::pin!(destroy_entered);
-        destroy_entered.as_mut().enable();
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-rejected")),
-        );
-
-        tokio::time::timeout(Duration::from_secs(5), destroy_entered)
-            .await
-            .expect("pool-full destroy should start");
-        assert_eq!(
-            hooks.destroy_count.load(Ordering::SeqCst),
-            1,
-            "rejected VM should be sent to destroy"
-        );
-        assert_eq!(
-            budget.allocated().2,
-            2,
-            "rejected active VM must retain its budget while destroy is in-flight"
-        );
-        {
-            let completions = env.handle.completions.lock().unwrap();
-            assert!(
-                !completions.iter().any(|c| c.run_id == run_id),
-                "provider.complete must wait until rejected VM destroy finishes"
-            );
-        }
-
-        hooks.release.notify_waiters();
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(c.is_some(), "job should complete after rejected VM destroy");
-        assert_eq!(c.unwrap().exit_code, 0);
-
-        wait_budget_count(&budget, 1, Duration::from_secs(2)).await;
-        let pool = idle_pool.lock().await;
-        assert_eq!(pool.len(), 1);
-        assert_eq!(pool.held_sessions(), vec!["sess-existing"]);
-        drop(pool);
-
-        shutdown(&env, run_handle).await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn parking_gate_closing_after_sandbox_park_rejects_and_waits_for_destroy() {
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::new());
-        let park_entered = Arc::new(tokio::sync::Notify::new());
-        let park_release = Arc::new(tokio::sync::Notify::new());
-        let destroy_entered = Arc::new(tokio::sync::Notify::new());
-        let destroy_release = Arc::new(tokio::sync::Notify::new());
-        overrides.set_park_gate(Arc::clone(&park_entered), Arc::clone(&park_release));
-        overrides.set_destroy_gate(Arc::clone(&destroy_entered), Arc::clone(&destroy_release));
-
-        let park_entered_wait = park_entered.notified();
-        tokio::pin!(park_entered_wait);
-        park_entered_wait.as_mut().enable();
-        let destroy_entered_wait = destroy_entered.notified();
-        tokio::pin!(destroy_entered_wait);
-        destroy_entered_wait.as_mut().enable();
-
-        let counter = Arc::clone(&overrides);
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 16384, 4, overrides);
-        let budget = Arc::clone(&config.budget);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-race-rejected")),
-        );
-
-        tokio::time::timeout(Duration::from_secs(5), park_entered_wait)
-            .await
-            .expect("sandbox park should start");
-        assert_eq!(counter.park_call_count(), 1);
-        assert_eq!(env.parking_gate.state(), ParkingState::Open);
-
-        env.drain();
-        assert_eq!(env.parking_gate.state(), ParkingState::SoftDraining);
-
-        park_release.notify_one();
-        tokio::time::timeout(Duration::from_secs(5), destroy_entered_wait)
-            .await
-            .expect("rejected parked sandbox should be destroyed");
-        assert_eq!(
-            counter.destroy_call_count(),
-            1,
-            "rejected VM should be sent to destroy exactly once"
-        );
-        assert_eq!(
-            budget.allocated().2,
-            1,
-            "rejected VM must retain budget while destroy is in-flight"
-        );
-        assert_eq!(
-            idle_pool.lock().await.len(),
-            0,
-            "closed gate must reject the candidate instead of parking it"
-        );
-        {
-            let completions = env.handle.completions.lock().unwrap();
-            assert!(
-                !completions.iter().any(|c| c.run_id == run_id),
-                "provider.complete must wait until rejected VM destroy finishes"
-            );
-        }
-
-        destroy_release.notify_one();
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        assert!(c.is_some(), "job should complete after destroy finishes");
-        assert_eq!(c.unwrap().exit_code, 0);
-
-        wait_budget_count(&budget, 0, Duration::from_secs(2)).await;
-        assert_eq!(idle_pool.lock().await.len(), 0);
-
-        shutdown(&env, run_handle).await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn cancellation_while_waiting_for_idle_pool_lock_destroys_instead_of_parking() {
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::new());
-        let park_entered = Arc::new(tokio::sync::Notify::new());
-        let park_release = Arc::new(tokio::sync::Notify::new());
-        let destroy_entered = Arc::new(tokio::sync::Notify::new());
-        let destroy_release = Arc::new(tokio::sync::Notify::new());
-        overrides.set_park_gate(Arc::clone(&park_entered), Arc::clone(&park_release));
-        overrides.set_destroy_gate(Arc::clone(&destroy_entered), Arc::clone(&destroy_release));
-
-        let park_entered_wait = park_entered.notified();
-        tokio::pin!(park_entered_wait);
-        park_entered_wait.as_mut().enable();
-        let destroy_entered_wait = destroy_entered.notified();
-        tokio::pin!(destroy_entered_wait);
-        destroy_entered_wait.as_mut().enable();
-
-        let counter = Arc::clone(&overrides);
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 16384, 4, overrides);
-        let budget = Arc::clone(&config.budget);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let cancel_tokens = Arc::clone(&config.cancel_tokens);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-cancel-while-locking")),
-        );
-        let token = wait_cancel_token(&cancel_tokens, run_id, Duration::from_secs(5)).await;
-
-        tokio::time::timeout(Duration::from_secs(5), park_entered_wait)
-            .await
-            .expect("sandbox park should start");
-        let pool_guard = idle_pool.lock().await;
-        park_release.notify_one();
-        // Let the job observe the first post-park cancel check, then block on
-        // the held pool lock. This opens the specific lock-wait window without
-        // relying on wall-clock sleeps.
-        tokio::task::yield_now().await;
-        token.cancel();
-        drop(pool_guard);
-
-        tokio::time::timeout(Duration::from_secs(5), destroy_entered_wait)
-            .await
-            .expect("cancelled lock-waiting sandbox should be destroyed");
-        assert_eq!(
-            counter.destroy_call_count(),
-            1,
-            "cancelled VM should be sent to destroy exactly once"
-        );
-        assert_eq!(
-            budget.allocated().2,
-            1,
-            "cancelled VM must retain budget while destroy is in-flight"
-        );
-        assert_eq!(
-            idle_pool.lock().await.len(),
-            0,
-            "cancelled VM must not enter the idle pool after waiting for the lock"
-        );
-
-        destroy_release.notify_one();
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await
-            .expect("job should complete after destroy finishes");
-        assert_eq!(c.exit_code, 0);
-        assert!(c.error.is_none());
-
-        wait_budget_count(&budget, 0, Duration::from_secs(2)).await;
-        wait_cancel_token_removed(&cancel_tokens, run_id, Duration::from_secs(2)).await;
-        assert_eq!(idle_pool.lock().await.len(), 0);
-
-        shutdown(&env, run_handle).await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn cancellation_during_sandbox_park_destroys_instead_of_parking() {
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::new());
-        let park_entered = Arc::new(tokio::sync::Notify::new());
-        let park_release = Arc::new(tokio::sync::Notify::new());
-        let destroy_entered = Arc::new(tokio::sync::Notify::new());
-        let destroy_release = Arc::new(tokio::sync::Notify::new());
-        overrides.set_park_gate(Arc::clone(&park_entered), Arc::clone(&park_release));
-        overrides.set_destroy_gate(Arc::clone(&destroy_entered), Arc::clone(&destroy_release));
-
-        let park_entered_wait = park_entered.notified();
-        tokio::pin!(park_entered_wait);
-        park_entered_wait.as_mut().enable();
-        let destroy_entered_wait = destroy_entered.notified();
-        tokio::pin!(destroy_entered_wait);
-        destroy_entered_wait.as_mut().enable();
-
-        let counter = Arc::clone(&overrides);
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 16384, 4, overrides);
-        let budget = Arc::clone(&config.budget);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let cancel_tokens = Arc::clone(&config.cancel_tokens);
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-cancel-while-parking")),
-        );
-        let token = wait_cancel_token(&cancel_tokens, run_id, Duration::from_secs(5)).await;
-
-        tokio::time::timeout(Duration::from_secs(5), park_entered_wait)
-            .await
-            .expect("sandbox park should start");
-        assert_eq!(counter.park_call_count(), 1);
-
-        token.cancel();
-        park_release.notify_one();
-
-        tokio::time::timeout(Duration::from_secs(5), destroy_entered_wait)
-            .await
-            .expect("cancelled parked sandbox should be destroyed");
-        assert_eq!(
-            counter.destroy_call_count(),
-            1,
-            "cancelled VM should be sent to destroy exactly once"
-        );
-        assert_eq!(
-            budget.allocated().2,
-            1,
-            "cancelled VM must retain budget while destroy is in-flight"
-        );
-        assert_eq!(
-            idle_pool.lock().await.len(),
-            0,
-            "cancelled VM must not enter the idle pool after park returns"
-        );
-        {
-            let completions = env.handle.completions.lock().unwrap();
-            assert!(
-                !completions.iter().any(|c| c.run_id == run_id),
-                "provider.complete must wait until cancelled VM destroy finishes"
-            );
-        }
-
-        destroy_release.notify_one();
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        let c = c.expect("job should complete after destroy finishes");
-        assert_eq!(c.exit_code, 0);
-        assert!(
-            c.error.is_none(),
-            "late cleanup cancellation should not rewrite job result"
-        );
-
-        wait_budget_count(&budget, 0, Duration::from_secs(2)).await;
-        wait_cancel_token_removed(&cancel_tokens, run_id, Duration::from_secs(2)).await;
-        assert_eq!(idle_pool.lock().await.len(), 0);
-
-        shutdown(&env, run_handle).await;
-    }
-
-    /// When the runner takes a sandbox out of the idle pool for reuse and
-    /// `Sandbox::unpark()` returns an error, the idle entry is destroyed
-    /// and the runner falls through to a fresh sandbox create.
-    #[tokio::test(start_paused = true)]
-    async fn unpark_failure_destroys_idle_entry_and_falls_through() {
-        // Both the pre-seeded sandbox and the fresh-create sandbox share
-        // the same MockSandboxOverrides set, so the unpark error queued
-        // here is consumed by the FIRST unpark call — which is the take
-        // path's call against the pre-seeded sandbox.
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::new());
-        let counter = Arc::clone(&overrides);
-        overrides.push_unpark_result(Err(sandbox::SandboxError::IdleTransition {
-            transition: sandbox::SandboxIdleTransition::Unpark,
-            message: "simulated unpark failure".into(),
-        }));
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 16384, 4, overrides);
-        let budget = Arc::clone(&config.budget);
-        let idle_pool = Arc::clone(&config.idle_pool);
-
-        // Pre-seed via the factory so the seeded MockSandbox shares the
-        // override set (and consumes the queued unpark error).
-        seed_idle_pool_with_overrides(
-            &idle_pool,
-            &budget,
-            &counter,
-            "sess-unpark-fail",
-            "vm0/default",
-            2,
-            4096,
-        )
-        .await;
-        assert_eq!(idle_pool.lock().await.len(), 1, "pool seeded");
-
-        let run_handle = tokio::spawn(run(config));
-
-        // Push a job for the same session — runner will try to reuse,
-        // unpark() will fail, idle entry gets destroyed, fresh create runs.
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-unpark-fail")),
-        );
-
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await;
-        let c = c.expect("fresh-create job should still complete");
-        assert_eq!(c.exit_code, 0);
-        assert_eq!(
-            c.reuse_result,
-            Some(SandboxReuseResult::UnparkFailed),
-            "completion must tag the unpark-failure branch",
-        );
-
-        // After the dust settles:
-        //   - unpark called exactly once (the failed take-side call);
-        //   - park called exactly once (the fresh-create's successful park).
-        tokio::time::sleep(Duration::from_millis(100)).await;
-        assert_eq!(
-            counter.unpark_call_count(),
-            1,
-            "expected exactly one unpark attempt"
-        );
-        assert_eq!(
-            counter.park_call_count(),
-            1,
-            "expected exactly one park (the fresh-create's)"
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn unpark_failure_status_switches_from_idle_to_active_while_job_runs() {
-        let gate = Arc::new(tokio::sync::Notify::new());
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::with_wait_exit_gate(
-            Arc::clone(&gate),
-        ));
-        let counter = Arc::clone(&overrides);
-        overrides.push_unpark_result(Err(sandbox::SandboxError::IdleTransition {
-            transition: sandbox::SandboxIdleTransition::Unpark,
-            message: "simulated unpark failure".into(),
-        }));
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 16384, 4, overrides);
-        let budget = Arc::clone(&config.budget);
-        let idle_pool = Arc::clone(&config.idle_pool);
-        let status = Arc::clone(&config.status);
-        let status_path = env._temp_dir.path().join("status.json");
-
-        seed_idle_pool_with_overrides(
-            &idle_pool,
-            &budget,
-            &counter,
-            "sess-unpark-status",
-            "vm0/default",
-            2,
-            4096,
-        )
-        .await;
-        publish_idle_status(&idle_pool, &status).await;
-        assert_eq!(
-            status_idle_sessions(&status_path).await,
-            vec!["sess-unpark-status".to_string()],
-            "pre-run status should list the idle VM",
-        );
-
-        let run_handle = tokio::spawn(run(config));
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-unpark-status")),
-        );
-
-        wait_idle_pool_len(&idle_pool, 0, Duration::from_secs(5)).await;
-        wait_status_idle_empty_with_active_run(&status_path, run_id, Duration::from_secs(5)).await;
-
-        gate.notify_one();
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await
-            .expect("fresh-create job should still complete");
-        assert_eq!(
-            completion.reuse_result,
-            Some(SandboxReuseResult::UnparkFailed),
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn unpark_panic_destroys_idle_entry_and_falls_through() {
-        let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::new());
-        let counter = Arc::clone(&overrides);
-        overrides.push_unpark_panic("simulated unpark panic");
-        let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 16384, 4, overrides);
-        let budget = Arc::clone(&config.budget);
-        let idle_pool = Arc::clone(&config.idle_pool);
-
-        seed_idle_pool_with_overrides(
-            &idle_pool,
-            &budget,
-            &counter,
-            "sess-unpark-panic",
-            "vm0/default",
-            2,
-            4096,
-        )
-        .await;
-
-        let run_handle = tokio::spawn(run(config));
-
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-unpark-panic")),
-        );
-
-        let c = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await
-            .expect("fresh-create job should still complete");
-        assert_eq!(c.exit_code, 0);
-        assert_eq!(c.reuse_result, Some(SandboxReuseResult::UnparkFailed));
-
-        wait_budget_count(&budget, 1, Duration::from_secs(2)).await;
-        assert_eq!(counter.unpark_call_count(), 1);
-        assert_eq!(counter.park_call_count(), 1);
-        assert_eq!(idle_pool.lock().await.len(), 1);
-
-        shutdown(&env, run_handle).await;
-    }
-
-    // -----------------------------------------------------------------------
-    // Reuse-enabled job whose session has no idle entry reports PoolMiss
-    // -----------------------------------------------------------------------
-
-    #[tokio::test(start_paused = true)]
-    async fn reuse_enabled_empty_pool_reports_pool_miss() {
-        let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
-        let idle_pool = Arc::clone(&config.idle_pool);
-
-        let run_handle = tokio::spawn(run(config));
-
-        // Empty pool + resume_session set + feature on → PoolMiss branch.
-        let run_id = RunId::new_v4();
-        push_job(
-            &env,
-            run_id,
-            "vm0/default",
-            Some(context_with_session(run_id, "sess-missing")),
-        );
-
-        let completion = env
-            .handle
-            .wait_completion(run_id, Duration::from_secs(5))
-            .await
-            .expect("job should complete");
-        assert_eq!(completion.exit_code, 0);
-        assert_eq!(
-            completion.reuse_result,
-            Some(SandboxReuseResult::PoolMiss),
-            "empty-pool reuse attempt must tag PoolMiss",
-        );
-        assert!(
-            completion.sandbox_id.is_some(),
-            "fresh create still allocates a sandbox id",
-        );
-        // Sanity: no one was in the pool to begin with.
-        assert_eq!(
-            idle_pool.lock().await.len(),
-            1,
-            "fresh-create sandbox re-parks into the pool",
-        );
-
-        shutdown(&env, run_handle).await;
-    }
-}
+mod tests;
