+++ crates/runner/src/cmd/start/tests/mod.rs
@@ -1,0 +1,6 @@
+mod budget;
+mod failure_panic;
+mod idle_reuse;
+mod main_loop;
+mod support;
+mod teardown_timer;
